Index

Exported

VoronoiFVM — Module

VoronoiFVM

VoronoiFVM.jl

Solver for coupled nonlinear partial differential equations (elliptic-parabolic conservation laws) based on the Voronoi finite volume method. It uses automatic differentiation via ForwardDiff.jl and DiffResults.jl to evaluate user functions along with their jacobians and calculate derivatives of solutions with respect to their parameters.

JuliaCon 2024 Lightning Talk: abstract, video

Recent changes

Please look up the list of recent changes

Accompanying packages

ExtendableSparse.jl: convenient and efficient sparse matrix assembly

ExtendableGrids.jl: unstructured grid management library

SimplexGridFactory.jl: unified high level mesh generator interface

Triangulate.jl: Julia wrapper for the Triangle triangle mesh generator by J. Shewchuk

TetGen.jl: Julia wrapper for the TetGen tetrahedral mesh generator by H. Si.

GridVisualize.jl: grid and function visualization related to ExtendableGrids.jl

PlutoVista.jl: backend for GridVisualize.jl for use in Pluto notebooks.

VoronoiFVM.jl and most of these packages are part of the meta package PDELib.jl.

Some alternatives

ExtendableFEM.jl: finite element library implementing gradient robust FEM from the same package base by Ch. Merdon

SkeelBerzins.jl: a Julian variation on Matlab's pdepe API

Trixi.jl: numerical simulation framework for hyperbolic conservation laws

GridAP.jl Grid-based approximation of partial differential equations in Julia

Ferrite.jl Finite element toolbox for Julia

FinEtools.jl Finite element tools for Julia

FiniteVolumeMethod.jl Finite volumes with Donald boxes

Some projects and packages using VoronoiFVM.jl

RfbScFVM: Performance prediction of flow battery vells

ChargeTransport.jl: Drift diffusion simulator for semiconductor devices

MosLab.jl: From semiconductor to transistor level modeling in Julia

LiquidElectrolytes.jl: Generalized Nernst-Planck-Poisson model for liquid electrolytes

Citation

If you use this package in your work, please cite it according to CITATION.cff

Types and Constructors

VoronoiFVM.AbstractAssemblyData

VoronoiFVM.AbstractData

VoronoiFVM.AbstractEdge

VoronoiFVM.AbstractEdgeData

VoronoiFVM.AbstractEvaluator

VoronoiFVM.AbstractGeometryItem

VoronoiFVM.AbstractImpedanceSystem

VoronoiFVM.AbstractNode

VoronoiFVM.AbstractNodeData

VoronoiFVM.AbstractPhysics

VoronoiFVM.AbstractQuantity

VoronoiFVM.AbstractSolutionArray

VoronoiFVM.AbstractSystem

VoronoiFVM.AbstractTransientSolution

VoronoiFVM.AssemblyError

VoronoiFVM.BEdge

VoronoiFVM.BICGstabIteration

VoronoiFVM.BNode

VoronoiFVM.BlockStrategy

VoronoiFVM.CGIteration

VoronoiFVM.CellwiseAssemblyData

VoronoiFVM.ContinuousQuantity

VoronoiFVM.ContinuousQuantity

VoronoiFVM.ConvergenceError

VoronoiFVM.DenseSolutionArray

VoronoiFVM.DenseSolutionArray

VoronoiFVM.DenseSolutionArray

VoronoiFVM.DenseSolutionArray

VoronoiFVM.DenseSystem

VoronoiFVM.DiffEqHistory

VoronoiFVM.DirectSolver

VoronoiFVM.DiscontinuousQuantity

VoronoiFVM.DiscontinuousQuantity

VoronoiFVM.Edge

VoronoiFVM.EdgewiseAssemblyData

VoronoiFVM.EmbeddingError

VoronoiFVM.EquationBlock

VoronoiFVM.Equationwise

VoronoiFVM.GMRESIteration

VoronoiFVM.ImpedanceSystem

VoronoiFVM.ImpedanceSystem

VoronoiFVM.ImpedanceSystem

VoronoiFVM.InterfaceQuantity

VoronoiFVM.InterfaceQuantity

VoronoiFVM.LinearSolverError

VoronoiFVM.LinearSolverStrategy

VoronoiFVM.NewtonControl

VoronoiFVM.NewtonSolverHistory

VoronoiFVM.NoBlock

VoronoiFVM.Node

VoronoiFVM.NodeRHS

VoronoiFVM.NodeUnknowns

VoronoiFVM.Physics

VoronoiFVM.Physics

VoronoiFVM.PointBlock

VoronoiFVM.ResEvaluator

VoronoiFVM.ResEvaluator

VoronoiFVM.ResJacEvaluator

VoronoiFVM.ResJacEvaluator

VoronoiFVM.SolverControl

VoronoiFVM.SparseSolutionArray

VoronoiFVM.SparseSolutionArray

VoronoiFVM.SparseSolutionIndices

VoronoiFVM.SparseSystem

VoronoiFVM.System

VoronoiFVM.System

VoronoiFVM.System

VoronoiFVM.System

VoronoiFVM.System

VoronoiFVM.SystemState

VoronoiFVM.SystemState

VoronoiFVM.SystemState

VoronoiFVM.TestFunctionFactory

VoronoiFVM.TestFunctionFactory

VoronoiFVM.TransientSolution

VoronoiFVM.TransientSolution

VoronoiFVM.TransientSolverHistory

VoronoiFVM.VectorOfDiskArrays

Constants

VoronoiFVM.sysmutatelock

Methods

VoronoiFVM.Grid

VoronoiFVM._add

VoronoiFVM._add

VoronoiFVM._add

VoronoiFVM._addnz

VoronoiFVM._complete!

VoronoiFVM._eval_and_assemble_generic_operator

VoronoiFVM._eval_res_jac!

VoronoiFVM._fill!

VoronoiFVM._print_error

VoronoiFVM._set

VoronoiFVM._set

VoronoiFVM.addzrows

VoronoiFVM.assemble_res

VoronoiFVM.assemble_res_jac

VoronoiFVM.bernoulli_horner

VoronoiFVM.bfacenodefactors

VoronoiFVM.bfacevelocities

VoronoiFVM.boundary_dirichlet!

VoronoiFVM.boundary_dirichlet!

VoronoiFVM.boundary_dirichlet!

VoronoiFVM.boundary_dirichlet!

VoronoiFVM.boundary_dirichlet!

VoronoiFVM.boundary_neumann!

VoronoiFVM.boundary_neumann!

VoronoiFVM.boundary_neumann!

VoronoiFVM.boundary_neumann!

VoronoiFVM.boundary_robin!

VoronoiFVM.boundary_robin!

VoronoiFVM.boundary_robin!

VoronoiFVM.boundary_robin!

VoronoiFVM.calc_divergences

VoronoiFVM.cartesian!

VoronoiFVM.circular_symmetric!

VoronoiFVM.coordinates

VoronoiFVM.data

VoronoiFVM.details

VoronoiFVM.dof

VoronoiFVM.dof

VoronoiFVM.doolittle_ludecomp!

VoronoiFVM.doolittle_lusolve!

VoronoiFVM.edgebatch

VoronoiFVM.edgeintegrate

VoronoiFVM.edgelength

VoronoiFVM.edgerange

VoronoiFVM.edgevelocities

VoronoiFVM.embedparam

VoronoiFVM.enable_boundary_species!

VoronoiFVM.enable_species!

VoronoiFVM.enable_species!

VoronoiFVM.eval_and_assemble

VoronoiFVM.eval_jacobian!

VoronoiFVM.eval_rhs!

VoronoiFVM.evaluate!

VoronoiFVM.evaluate!

VoronoiFVM.evaluate!

VoronoiFVM.evaluate_residual_and_jacobian

VoronoiFVM.factorizationstrategy

VoronoiFVM.fbernoulli

VoronoiFVM.fbernoulli_pm

VoronoiFVM.firstnodedof

VoronoiFVM.fixed_timesteps!

VoronoiFVM.freqdomain_impedance

VoronoiFVM.generic_operator_sparsity!

VoronoiFVM.getdof

VoronoiFVM.getnodedof

VoronoiFVM.getspecies

VoronoiFVM.h1norm

VoronoiFVM.h1seminorm

VoronoiFVM.hasdata

VoronoiFVM.hasoutflownode

VoronoiFVM.history

VoronoiFVM.history_details

VoronoiFVM.history_summary

VoronoiFVM.impedance

VoronoiFVM.increase_num_species!

VoronoiFVM.inplace_linsolve!

VoronoiFVM.inplace_linsolve!

VoronoiFVM.integrate

VoronoiFVM.integrate

VoronoiFVM.integrate

VoronoiFVM.integrate_stdy

VoronoiFVM.integrate_tran

VoronoiFVM.is_boundary_species

VoronoiFVM.is_bulk_species

VoronoiFVM.isnodespecies

VoronoiFVM.isnontrivial

VoronoiFVM.isoutflownode

VoronoiFVM.isregionspecies

VoronoiFVM.isunknownsof

VoronoiFVM.jac

VoronoiFVM.l2h1norm

VoronoiFVM.l2h1seminorm

VoronoiFVM.l2norm

VoronoiFVM.lastnodedof

VoronoiFVM.lpnorm

VoronoiFVM.lpw1pnorm

VoronoiFVM.lpw1pseminorm

VoronoiFVM.mass_matrix

VoronoiFVM.meas

VoronoiFVM.measurement_derivative

VoronoiFVM.nodebatch

VoronoiFVM.nodeflux

VoronoiFVM.noderange

VoronoiFVM.nodevolumes

VoronoiFVM.nondelaunay

VoronoiFVM.num_dof

VoronoiFVM.num_quantities

VoronoiFVM.num_species

VoronoiFVM.outflownode

VoronoiFVM.outflownode!

VoronoiFVM.parameters

VoronoiFVM.partitioning

VoronoiFVM.physics!

VoronoiFVM.plothistory

VoronoiFVM.prepare_diffeq!

VoronoiFVM.project

VoronoiFVM.ramp

VoronoiFVM.region

VoronoiFVM.res

VoronoiFVM.res

VoronoiFVM.setdof!

VoronoiFVM.solve_step!

VoronoiFVM.solve_transient!

VoronoiFVM.spherical_symmetric!

VoronoiFVM.subgrids

VoronoiFVM.subgrids

VoronoiFVM.testfunction

VoronoiFVM.time

VoronoiFVM.unknown_indices

VoronoiFVM.unknown_indices

VoronoiFVM.unknowns

VoronoiFVM.unknowns

VoronoiFVM.unknowns

VoronoiFVM.update_grid!

VoronoiFVM.update_grid_cellwise!

VoronoiFVM.update_grid_edgewise!

VoronoiFVM.value

VoronoiFVM.values

VoronoiFVM.values

VoronoiFVM.viewK

VoronoiFVM.viewL

VoronoiFVM.views

VoronoiFVM.w1pnorm

VoronoiFVM.w1pseminorm