Grid

Grid

Types and Constants

VoronoiFVM.AbstractGrid — Type.

abstract type AbstractGrid

Abstract type for grid like datastructures VoronoiFVM.Grid and VoronoiFVM.SubGrid.

source

VoronoiFVM.Grid — Type.

struct Grid{Tc} <: VoronoiFVM.AbstractGrid

Structure holding grid data. It is parametrised by the type Tc of coordinates.

coord::Array{Tc,2} where Tc

2D Array of node coordinates

cellnodes::Array{Int32,2}

2D Array of node indices per grid cell

cellregions::Array{Int32,1}

Array of cell region numbers

bfacenodes::ElasticArrays.ElasticArray{Int32,2,1}

2D Array of node indices per boundary face

bfaceregions::Array{Int32,1}

Array of boundary face region numbers

num_cellregions::Array{Int32,1}

Number of inner cell regions. Stored in an array in order to keep the struct immutable.

num_bfaceregions::Array{Int32,1}

Number of boundary face regions. Stored in an array in order to keep the struct immutable.

celledgenodes::Array{Int32,2}

2D Array describing local scheme of distributions nodes per cell edge.

source

VoronoiFVM.Grid — Method.

Constructor for 1D grid.

Construct 1D grid from an array of node cordinates. It creates two boundary regions with index 1 at the left end and index 2 at the right end.

Primal grid holding unknowns: marked by o, dual grid marking control volumes: marked by |.

 o-----o-----o-----o-----o-----o-----o-----o-----o
 |--|-----|-----|-----|-----|-----|-----|-----|--|

source

VoronoiFVM.Grid — Method.

Grid(X, Y)

Constructor for 2D grid from coordinate arrays. Boundary region numbers count counterclockwise:

location	number

south	1

east	2

north	3

west	4

source

VoronoiFVM.SubGrid — Type.

struct SubGrid{Tc} <: VoronoiFVM.AbstractGrid

Subgrid of parent grid (mainly for visualization purposes). Intended to hold support of species which are not defined everywhere.

parent::VoronoiFVM.Grid

Parent Grid

node_in_parent::Array{Int32,1}

Incidence between subgrid node numbers and node numbers in parent.

coord::Array{Tc,2} where Tc

2D Array of coordinates per grid node

cellnodes::Array{Int32,2}

2D Array of node numbers per grid cell

source

Methods

VoronoiFVM.bfacemask! — Method.

bfacemask!(grid::VoronoiFVM.Grid, maskmin::AbstractArray, maskmax::AbstractArray, ireg::Int64; tol) -> Int64

Edit region numbers of grid boundary facets via rectangular mask. Currently, only for 1D grids, inner boundaries can be added.

source

VoronoiFVM.cellmask! — Method.

cellmask!(grid::VoronoiFVM.Grid, maskmin::AbstractArray, maskmax::AbstractArray, ireg::Int64; tol) -> Int64

Edit region numbers of grid cells via rectangular mask.

source

VoronoiFVM.num_bfaces — Method.

num_bfaces(grid::VoronoiFVM.Grid) -> Int64

Number of boundary faces in grid.

source

VoronoiFVM.num_cells — Method.

num_cells(grid::VoronoiFVM.AbstractGrid) -> Any

Number of cells in grid

source

VoronoiFVM.num_nodes — Method.

num_nodes(grid::VoronoiFVM.AbstractGrid) -> Any

Number of nodes in grid

source

VoronoiFVM.subgrid — Method.

subgrid(parent::VoronoiFVM.Grid, subregions::AbstractArray; transform, boundary) -> VoronoiFVM.SubGrid{_1} where _1

Create subgrid of list of regions.

source

VoronoiFVM.glue — Method.

Glue together two vectors a and b resulting in a vector c. They last element of a shall be equal (up to tol) to the first element of b. The result fulfills length(c)=length(a)+length(b)-1

source

VoronoiFVM.value — Function.

Extract value from dual number. Use to debug physics callbacks. Re-exported from ForwardDiff.jl

source

Base.eltype — Method.

eltype(grid::VoronoiFVM.AbstractGrid) -> Any

Return element type of grid coordinates.

source

VoronoiFVM.bfacefactors! — Method.

bfacefactors!(grid, icell, nodefac)

Calculate node volume and voronoi surface contributions for boundary face.

source

VoronoiFVM.bfacenode — Method.

bfacenode(grid::VoronoiFVM.Grid, inode::Any, icell::Any) -> Any

Index of boundary face node.

source

VoronoiFVM.celledgenode — Method.

celledgenode(grid::VoronoiFVM.Grid, inode::Any, iedge::Any, icell::Any) -> Any

Index of cell edge node.

source

VoronoiFVM.cellfactors! — Method.

cellfactors!(grid, icell, nodefac, edgefac)

Calculate node volume and voronoi surface contributions for cell.

source

VoronoiFVM.cellnode — Method.

cellnode(grid::VoronoiFVM.AbstractGrid, inode::Any, icell::Any) -> Any

Return index of i-th local node in cell icell

source

VoronoiFVM.dim_grid — Method.

dim_grid(grid::VoronoiFVM.Grid) -> Int64

Topological dimension of grid

source

VoronoiFVM.dim_space — Method.

dim_space(grid::VoronoiFVM.AbstractGrid) -> Any

Space dimension of grid

source

VoronoiFVM.nodecoord — Method.

nodecoord(grid::VoronoiFVM.AbstractGrid, inode::Any) -> SubArray

Return view of coordinates of node inode.

source

VoronoiFVM.num_bfaceregions — Method.

num_bfaceregions(grid::VoronoiFVM.Grid) -> Int32

Number of boundary face regions in grid.

source

VoronoiFVM.num_cellregions — Method.

num_cellregions(grid::VoronoiFVM.Grid) -> Int32

Number of cell regions in grid.

source

VoronoiFVM.num_edges_per_cell — Method.

num_edges_per_cell(grid::VoronoiFVM.Grid) -> Int64

Number of edges per grid cell.

source

VoronoiFVM.num_nodes_per_bface — Method.

num_nodes_per_bface(grid::VoronoiFVM.Grid) -> Int64

Number of nodes per boundary face

source

VoronoiFVM.num_nodes_per_cell — Method.

num_nodes_per_cell(grid::VoronoiFVM.AbstractGrid) -> Any

Return number of nodes per cell in grid.

source

VoronoiFVM.reg_bface — Method.

reg_bface(grid::VoronoiFVM.Grid, icell::Any) -> Any

Boundary region number for boundary face

source

VoronoiFVM.reg_cell — Method.

reg_cell(grid::VoronoiFVM.Grid, icell::Any) -> Any

Bulk region number for cell

source

VoronoiFVM.geomspace — Method.

geomspace(a, b, ha, hb; tol)

(Try to) create a subdivision of interval (a,b) stored in the returned array X such that

X[1]==a, X[end]==b

(X[2]-X[1])<=ha+tol*(b-a)

(X[end]-X[end-1])<=hb+tol*(b-a)

There is a number q such that X[i+1]-X[i] == q*(X[i]-X[i-1])

X is the array with the minimal possible number of points with the above property

Caveat: the algorithm behind this is well tested but unproven.

Returns an Array containing the points of the subdivision.

source

VoronoiFVM.tricircumcenter! — Method.

tricircumcenter!(circumcenter, a, b, c)

Find the circumcenter of a triangle.

The result is returned both in terms of x-y coordinates and xi-eta coordinates, relative to the triangle's point a' (that is,a' is the origin of both coordinate systems). Hence, the x-y coordinates returned are NOT absolute; one must add the coordinates of `a' to find the absolute coordinates of the circumcircle. However, this means that the result is frequently more accurate than would be possible if absolute coordinates were returned, due to limited floating-point precision. In general, the circumradius can be computed much more accurately.

Created from C source of Jonathan R Shewchuk <jrs@cs.cmu.edu>

Modified to return absolute coordinates.

source