OrdinaryDiffEq.jl solver

For transient problems, as an alternative to the use of the built-in implicit Euler method, (stiff) ODE solvers from OrdinaryDiffEq.jl can be used.

The interface just provides two methods: creation of an ODEProblem from a VoronoiFVM.System and a reshape method which turns the output of the ode solver into a TransientSolution.

The basic usage pattern is as follows: use OrdinaryDiffEq.jl and replace the call to the built-in solver

sol=solve(sys; times=(t0,t1), inival=inival)

by

problem = ODEProblem(sys,inival,(t0,tend))
odesol = solve(problem, solver)
sol=reshape(odesol,sys)

Here, solver is one of the aforementioned ODE solvers. If the DifferentialEquations.jl package is used, the solver parameter is omitted, and some default is chosen.

SciMLBase.ODEFunction — Method

 ODEFunction(system,inival=unknowns(system,inival=0),t0=0)

Create an ODEPFunction in mass matrix form to be handeled by ODE solvers from DifferentialEquations.jl.

Parameters:

system: A VoronoiFVM.System

jacval (optional): Initial value. Default is a zero vector. Consider to pass a stationary solution at time tjac.

tjac (optional): tjac, Default: 0

The jacval and tjac are passed for a first evaluation of the Jacobian, allowing to detect the sparsity pattern which is passed to the solver.

SciMLBase.ODEProblem — Method

ODEProblem(system,inival,tspan,callback=SciMLBase.CallbackSet())

Create an ODEProblem in mass matrix form which can be handeled by ODE solvers from DifferentialEquations.jl.

Parameters:

system: A VoronoiFVM.System

inival: Initial value. Consider to pass a stationary solution at tspan[1].

tspan: Time interval

callback : (optional) callback for ODE solver

The method returns an ODEProblem which can be solved by solve().

The solution sol returned by solve conforms to the ArrayInterface but "forgot" the VoronoiFVM species structure. Accessing sol(t) will return an interpolated solution vector giving the value of the solution at moment t. Using reshape(sol(t),system) it can be turned into into a sparse or dense array reflecting the species structure of system. The order of the interpolation depends on the ODE solver.

Base.reshape — Method

reshape(v, system)

Reshape vector to fit as solution to system.

reshape(ode_solution, system; times)

Create a TransientSolution from the output of the ode solver which reflects the species structure of the probem which is not seen by the ODE solver, howvever the interpolation behind reshaped_sol(t) will be linear.

If times is specified, the interpolated solution at the given moments of time will be returned.

Choice of ODE/DAE solver

As this package interfaces to the PDE solver package VoronoiFVM.jl, a general advise is to choose methods able to handle stiff problems. Moreover, often, discretized PDE systems (e.g. containing elliptic equations) are differential agebraic equation (DAE) systems which should be solved by DAE solvers.