Plotting

In order to avoid drawing in of heavy dependencies for VoronoiFVM, the plot methods for grids and grid functions defined in this package have as their first argument the module of the plotting package used.

Currently, PyPlot and Plots are supported. Similar schemes in the future might work for Makie and other packages.

VoronoiFVM.isplots — Method

isplots(Plotter::Any) -> Bool

Check if plotter is Plots

VoronoiFVM.ispyplot — Method

ispyplot(Plotter::Any) -> Bool

Check if plotter is PyPlot

VoronoiFVM.frgb — Method

frgb(Plotter::Any, i::Any, max::Any; pastel) -> Any

Plot color scale for grid colors.

VoronoiFVM.isplotter — Method

isplotter(Plotter::Any) -> Bool

Check if plotter is a plotter at all.

VoronoiFVM.plot — Method

plot(Plotter::Any, grid::VoronoiFVM.AbstractGrid, U::AbstractArray; color, cmap, label, levels, aspect, clear, show) -> Any

Plot array as piecewise linear function on grid.

VoronoiFVM.plot — Method

plot(Plotter::Any, grid::VoronoiFVM.Grid; aspect, clear, show, p) -> Any

Plot grid.

VoronoiFVM.plot — Method

plot(Plotter::Any, grid::VoronoiFVM.SubGrid, U::AbstractArray; color, label, clear, show, p) -> Any

Plot array as piecewise constant function on subgrid

VoronoiFVM.plot — Method

plot(Plotter::Any, subgrid::VoronoiFVM.SubGrid; clear, show) -> Any

Plot subgrid.