Interface to DifferentialEquations.jl

VoronoiFVM.eval_jacobian! — Method

eval_jacobian!(J, u, sys, t)

Interpret the discrete problem as an ODE/DAE problem. Provide the jacobi matrix calculation function for DifferentialEquations.jl.

VoronoiFVM.eval_rhs! — Method

eval_rhs!(du, u, sys, t)

Interpret the discrete problem as an ODE/DAE problem. Provide the rhs function for DifferentialEquations.jl.

VoronoiFVM.jac_prototype — Method

jac_prototype(sys)

Complete the system and provide the jacobi matrix as prototype for the jacobian for use with DifferentialEquations.jl.

VoronoiFVM.mass_matrix — Method

mass_matrix(system)

Calculate the mass matrix for use with DifferentialEquations.jl.

VoronoiFVM.solve — Method

solve(DifferentialEquations, inival, system, tspan; 
      solver=nothing, 
      kwargs...)

Use a timesteping scheme from DifferentialEquations.jl to perform transient solution of the system.

The system must have a constant diagonal mass matrix.

Any stiff solver capable to work with mass matrices is possible. If the system contains elliptic equations, it needs to be a DAE solver able to work with. See a list here.

The constructed solver should be passed as solver= kwarg. By default (if nothing is passed), the Rosenbrock23 solver is used.

All other kwargs are passed to DifferentialEquations.solve(), see here for an overview.

In order to avoid a package dependency on DifferentialEquations, the module corresponding to that package is passed as the first argument.

Returns a transient solution object sol containing stored solutions, see TransientSolution.