Grid

Types and Constants

VoronoiFVM.TokenStream — Type

mutable struct TokenStream

Tokenstream allows to read tokenized data from file without keeping the file ocntent in memory.

input::IOStream

Input stream

tokens::Array{SubString{String},1}

Array of current tokens kept in memory.

itoken::Int64

Position of actual token in tokens array

lineno::Int64

Line number in IOStream

comment::Char

Comment character

dlm::Function

Function telling if given character is a delimiter.

VoronoiFVM.TokenStream — Method

TokenStream(input::IOStream; comment, dlm) -> TokenStream

Create Tokenstream with IOStream argument.

VoronoiFVM.TokenStream — Method

TokenStream(filename::String; comment, dlm) -> TokenStream

Create Tokenstream with file name argument.

VoronoiFVM.Grid — Method

Grid(#temp#::Type{#s73} where #s73<:IOStream; file, format) -> VoronoiFVM.Grid{Float64,Int32}

Read grid from file.

VoronoiFVM.AbstractGrid — Type

abstract type AbstractGrid

Abstract type for grid like datastructures VoronoiFVM.Grid and VoronoiFVM.SubGrid.

VoronoiFVM.Cartesian1D — Type

struct Cartesian1D

Type for dispatching formfactor calculations. Intepret 1D coordinates as cartesian.

VoronoiFVM.Cartesian2D — Type

struct Cartesian2D

Type for dispatching formfactor calculations. Intepret 2D coordinates as cartesian.

VoronoiFVM.Cartesian3D — Type

struct Cartesian3D

Type for dispatching formfactor calculations. Intepret 3D coordinates as cartesian.

VoronoiFVM.CircularSymmetric1D — Type

struct CircularSymmetric1D

Type for dispatching formfactor calculations. Intepret 1D coordinates as radial coordinate r assuming circular symmetry around origin 0.

VoronoiFVM.CircularSymmetric2D — Type

struct CircularSymmetric2D

Type for dispatching formfactor calculations. Intepret 2D coordinates as radial coordinate r and eight coordinate z assuming circular symmetry around axis r=0.

VoronoiFVM.Grid — Type

mutable struct Grid{Tc, Ti} <: VoronoiFVM.AbstractGrid

Structure holding grid data. It is parametrised by the type Tc of coordinates.

coord::Array{Tc,2} where Tc

2D Array of node coordinates

cellnodes::ElasticArrays.ElasticArray{Ti,2,M} where M where Ti

2D Array of node indices per grid cell

celledges::Array{Ti,2} where Ti

Array of edge indices per grid cell. Instatiated with prepare_edges!(grid)

edgecells::Array{Ti,2} where Ti

Array of cell indices per grid edge. The number of these indices may be less than the number of columns in this array. Nonexisting indices are set to 0 Instatiated with prepare_edges!(grid)

edgenodes::Array{Ti,2} where Ti

Array of node indices per grid edge. Instatiated with prepare_edges!(grid)

cellregions::Array{Ti,1} where Ti

Array of cell region numbers

bfacenodes::ElasticArrays.ElasticArray{Ti,2,M} where M where Ti

2D Array of node indices per boundary face

bfaceregions::Array{Ti,1} where Ti

Array of boundary face region numbers

num_cellregions::Any

Number of inner cell regions.

num_bfaceregions::Any

Number of boundary face regions.

local_celledgenodes::Array{Ti,2} where Ti

2D Array describing local scheme of distributions nodes per cell edge.

coord_type::DataType

Type of coordinate system

VoronoiFVM.Grid — Method

Grid(coord, cellnodes, cellregions, bfacenodes, bfaceregions)

Main constructor for general grid from basic incidence information

VoronoiFVM.SphericalSymmetric1D — Type

struct SphericalSymmetric1D

Type for dispatching formfactor calculations. Intepret 1D coordinates as radial coordinate r assuming spehrical symmetry around point r=0.

VoronoiFVM.SubGrid — Type

struct SubGrid{Tc} <: VoronoiFVM.AbstractGrid

Subgrid of parent grid (mainly for visualization purposes). Intended to hold support of species which are not defined everywhere.

parent::VoronoiFVM.Grid

Parent Grid

node_in_parent::Array{Int32,1}

Incidence between subgrid node numbers and node numbers in parent.

coord::Array{Tc,2} where Tc

2D Array of coordinates per grid node

cellnodes::Array{Int32,2}

2D Array of node numbers per grid cell

VoronoiFVM.Grid — Method

Grid(X)

Constructor for 1D grid.

Construct 1D grid from an array of node cordinates. It creates two boundary regions with index 1 at the left end and index 2 at the right end.

Primal grid holding unknowns: marked by o, dual grid marking control volumes: marked by |.

 o-----o-----o-----o-----o-----o-----o-----o-----o
 |--|-----|-----|-----|-----|-----|-----|-----|--|

VoronoiFVM.Grid — Method

Grid(X, Y)

Constructor for 2D grid from coordinate arrays. Boundary region numbers count counterclockwise:

location	number

south	1

east	2

north	3

west	4

VoronoiFVM.Grid — Method

Grid(triangle_switches::String, tio_in::TriangulateIO) -> VoronoiFVM.Grid{Float64,Int32}

Create Grid from Triangle input data.

VoronoiFVM.Grid — Method

Grid(; flags, points, bfaces, bfaceregions, regionpoints, regionnumbers, regionvolumes) -> VoronoiFVM.Grid{Float64,Int32}

Create Grid from a number of input arrays. The 2D input arrays are transposed if necessary and converted to the proper data types for Triangulate.

This conversion is not performed if the data types are thos indicated in the defaults and the leading dimension of 2D arrays corresponds to the space dimension.

VoronoiFVM.UnexpectedTokenError — Type

struct UnexpectedTokenError <: Exception

Error thrown when the token expected in expect! is not there.

found::String

expected::String

lineno::Int64

Methods

VoronoiFVM.bfacenode — Method

bfacenode(grid::VoronoiFVM.Grid, inode::Any, icell::Any) -> Any

Index of boundary face node.

VoronoiFVM.cellnode — Method

cellnode(grid::VoronoiFVM.Grid, inode::Any, icell::Any) -> Any

Return index of i-th local node in cell icell

VoronoiFVM.dim_space — Method

dim_space(grid::VoronoiFVM.Grid) -> Int64

Space dimension of grid

VoronoiFVM.nodecoord — Method

nodecoord(grid::VoronoiFVM.Grid, inode::Any) -> SubArray

Return view of coordinates of node inode.

VoronoiFVM.num_bfaceregions — Method

num_bfaceregions(grid::VoronoiFVM.Grid) -> Any

Number of boundary face regions in grid.

VoronoiFVM.num_bfaces — Method

num_bfaces(grid::VoronoiFVM.Grid) -> Int64

Number of boundary faces in grid.

VoronoiFVM.num_cellregions — Method

num_cellregions(grid::VoronoiFVM.Grid) -> Any

Number of cell regions in grid.

VoronoiFVM.num_cells — Method

num_cells(grid::VoronoiFVM.Grid) -> Int64

Number of cells in grid

VoronoiFVM.num_edges — Method

num_edges(grid::VoronoiFVM.Grid) -> Int64

Number of edges in grid

VoronoiFVM.num_nodes — Method

num_nodes(grid::VoronoiFVM.Grid) -> Int64

Number of nodes in grid

VoronoiFVM.tridata — Method

tridata(g::Any) -> Tuple{Any,Any,Any}

Provide information for PyPlot triangulation data for plotting. Needs to be splatted. Use it e.g. like

    PyPlot.trisurf(tridata(g)...,U)

VoronoiFVM.cartesian! — Method

cartesian!(grid)

Set cartesian coordinates for grid.

VoronoiFVM.circular_symmetric! — Method

circular_symmetric!(grid)

Set circular coordinates for grid (1D or 2D).

VoronoiFVM.prepare_edges! — Method

prepare_edges!(grid)

Prepare edge adjacencies (celledges, edgecells, edgenodes)

VoronoiFVM.bfacemask! — Method

bfacemask!(grid::VoronoiFVM.Grid, maskmin::AbstractArray, maskmax::AbstractArray, ireg::Int64; tol) -> VoronoiFVM.Grid

Edit region numbers of grid boundary facets via rectangular mask. Currently, only for 1D grids, inner boundaries can be added.

VoronoiFVM.cellmask! — Method

cellmask!(grid::VoronoiFVM.Grid, maskmin::AbstractArray, maskmax::AbstractArray, ireg::Int64; tol) -> Any

Edit region numbers of grid cells via rectangular mask.

VoronoiFVM.cellnode — Method

cellnode(grid::VoronoiFVM.SubGrid, inode::Any, icell::Any) -> Any

Return index of i-th local node in cell icell

VoronoiFVM.dim_space — Method

dim_space(grid::VoronoiFVM.SubGrid) -> Int64

Space dimension of grid

VoronoiFVM.nodecoord — Method

nodecoord(grid::VoronoiFVM.SubGrid, inode::Any) -> SubArray

Return view of coordinates of node inode.

VoronoiFVM.num_cells — Method

num_cells(grid::VoronoiFVM.SubGrid) -> Int64

Number of cells in grid

VoronoiFVM.num_edges — Method

num_edges(grid::VoronoiFVM.SubGrid)

Number of edges in grid

VoronoiFVM.num_nodes — Method

num_nodes(grid::VoronoiFVM.SubGrid) -> Int64

Number of nodes in grid

VoronoiFVM.subgrid — Method

subgrid(parent::VoronoiFVM.Grid, subregions::AbstractArray; transform, boundary) -> VoronoiFVM.SubGrid{_1} where _1

Create subgrid of list of regions.

VoronoiFVM.geomspace — Method

geomspace(a, b, ha, hb; tol)

(Try to) create a subdivision of interval (a,b) stored in the returned array X such that

X[1]==a, X[end]==b

(X[2]-X[1])<=ha+tol*(b-a)

(X[end]-X[end-1])<=hb+tol*(b-a)

There is a number q such that X[i+1]-X[i] == q*(X[i]-X[i-1])

X is the array with the minimal possible number of points with the above property

Caveat: the algorithm behind this is well tested but unproven.

Returns an Array containing the points of the subdivision.

VoronoiFVM.glue — Method

glue(a, b; tol)

Glue together two vectors a and b resulting in a vector c. They last element of a shall be equal (up to tol) to the first element of b. The result fulfills length(c)=length(a)+length(b)-1

VoronoiFVM.expecttoken — Method

expecttoken(tks::TokenStream, expected::String) -> Bool

Expect keyword token.

If token is missing, an UnexpectedTokenError is thrown If the token has been found, reading will continue at the position after the token found.

VoronoiFVM.gettoken — Method

gettoken(tks::TokenStream) -> Union{Nothing, SubString{String}}

Get next token from tokenstream.

VoronoiFVM.trytoken — Method

trytoken(tks::TokenStream, expected::String) -> Bool

Try for keyword token.

It token is missing, the token read is put back into stream, a value of false is returned and the next try/gettoken command continues at the same position,

Otherwise, true is returned, and reading continues after the token found.

VoronoiFVM.edgevelocities — Method

edgevelocities(grid, velofunc)

Project velocity onto grid edges,

VoronoiFVM.bfacefactors! — Method

bfacefactors!(grid, icell, nodefac)

Calculate node volume contributions for boundary face.

VoronoiFVM.cellfactors! — Method

cellfactors!(grid, icell, nodefac, edgefac)

Calculate node volume and voronoi surface contributions for cell.

Base.eltype — Method

eltype(grid::VoronoiFVM.Grid) -> Any

Return element type of grid coordinates.

VoronoiFVM.celledge — Method

celledge(grid::VoronoiFVM.Grid, iedge::Any, icell::Any) -> Any

Index of cell edge node.

VoronoiFVM.celledgenode — Method

celledgenode(grid::VoronoiFVM.Grid, inode::Any, iedge::Any, icell::Any) -> Any

Index of cell edge node.

VoronoiFVM.dim_grid — Method

dim_grid(grid::VoronoiFVM.Grid) -> Int64

Topological dimension of grid

VoronoiFVM.num_edges_per_cell — Method

num_edges_per_cell(grid::VoronoiFVM.Grid) -> Int64

Number of edges per grid cell.

VoronoiFVM.num_nodes_per_bface — Method

num_nodes_per_bface(grid::VoronoiFVM.Grid) -> Int64

Number of nodes per boundary face

VoronoiFVM.num_nodes_per_cell — Method

num_nodes_per_cell(grid::VoronoiFVM.Grid) -> Int64

Return number of nodes per cell in grid.

VoronoiFVM.reg_bface — Method

reg_bface(grid::VoronoiFVM.Grid, icell::Any) -> Any

Boundary region number for boundary face

VoronoiFVM.reg_cell — Method

reg_cell(grid::VoronoiFVM.Grid, icell::Any) -> Any

Bulk region number for cell

VoronoiFVM.spherical_symmetric! — Method

spherical_symmetric!(grid)

Set spherical coordinates for grid (1D).

Base.eltype — Method

eltype(grid::VoronoiFVM.SubGrid) -> Any

Return element type of grid coordinates.

VoronoiFVM.num_nodes_per_cell — Method

num_nodes_per_cell(grid::VoronoiFVM.SubGrid) -> Int64

Return number of nodes per cell in grid.

Base.eof — Method

eof(tks::TokenStream) -> Bool

Check if all tokens have been consumed.

VoronoiFVM.destruct! — Method

destruct!(tks::TokenStream)

Tokenstream destructor should close input

VoronoiFVM.tricircumcenter! — Method

tricircumcenter!(circumcenter, a, b, c)

Find the circumcenter of a triangle.

Created from C source of Jonathan R Shewchuk <jrs@cs.cmu.edu>

Modified to return absolute coordinates.