Physics & special functions

Physics

VoronoiFVM.AbstractData — Type

abstract type AbstractData

Abstract type for user data.

Compat

Will be removed in future versions

VoronoiFVM.AbstractPhysics — Type

abstract type AbstractPhysics

Abstract type for physics.

VoronoiFVM.Physics — Type

struct Physics

Physics data record with the following fields:

flux::Function

Flux between neigboring control volumes: flux(f,_u,edge) or flux(f,_u,edge,data) should return in f[i] the flux of species i along the edge joining circumcenters of neigboring control volumes. u=unknowns(_u) returns a 2D array such that for species i, u[i,1] and u[i,2] contain the unknown values at the corresponding ends of the edge.

storage::Function

Storage term (term under time derivative): storage(f,u,node) or storage(f,u,node,data)

It should return in f[i] the storage term for the i-th equation. u[i] contains the value of the i-th unknown.

reaction::Function

Reaction term: reaction(f,u,node) or reaction(f,u,node,data)

It should return in f[i] the reaction term for the i-th equation. u[i] contains the value of the i-th unknown.

source::Function

Source term: source(f,node) or source(f,node,data).

It should return the in f[i] the value of the source term for the i-th equation.

bflux::Function

Flux between neighboring control volumes on the boundary

breaction::Function

Boundary reaction term: breaction(f,u,node) or breaction(f,u,node,data) Similar to reaction, but restricted to the inner or outer boundaries.

bsource::Function

Boundary source term: bsource(f,node) or bsource(f,node,data).

It should return in f[i] the value of the source term for the i-th equation.

bstorage::Function

Boundary storage term: bstorage(f,u,node) or bstorage(f,u,node,data) Similar to storage, but restricted to the inner or outer boundaries.

generic_operator::Function

Generic operator generic_operator(f,u,sys). This operator acts on the full solution u of a system. Sparsity is detected automatically unless generic_operator_sparsity is given.

generic_operator_sparsity::Function

Function defining the sparsity structure of the generic operator. This should return the sparsity pattern of the generic_operator.

data::Any

User data (parameters). This allows to pass various parameters to the callback functions.

num_species::Int8

Number of species including boundary species.

VoronoiFVM.Physics — Method

Physics(;num_species=1,
         data=nothing,
         flux,
         reaction,
         storage,
         source,
         breaction,
         bstorage,
         generic,
         generic_sparsity
)

Constructor for physics data. For the meaning of the optional function callbacks, see Physics.

There are two variants of this constructor. It data is given, all callback functions should accept a last data argument. Otherwise, no data are passed explicitely, and it is assumed that constitutive callbacks take parameters from the closure where the function is defined.

Base.show — Method

show(io, physics)

Show physics object

VoronoiFVM.hasdata — Method

hasdata(physics)

Check if physics object has data

VoronoiFVM.@create_physics_wrappers — Macro

@create_physics_wrappers(physics,node,bnode,edge)

Create wrapper functions around physics callbacks which fit the API of ForwardDiff.jacobian! and pass the data parameter if necessary. These are meant to be defined before performing assembly loops. The macro creates the follwing variables:

wrapper functions: fluxwrap,storagewrap,reactionwrap,bstoragewrap, breactionwrap

flag variables: issource, isreaction,isbreaction,isbstorage

Edge and node data

VoronoiFVM.BNode — Type

mutable struct BNode{Tv, Ti} <: VoronoiFVM.AbstractGeometryItem{Tv, Ti}

Structure holding local boundary node information.

index::Any

Index in grid

ibface::Any

BFace number it is called from

ibnode::Any

local node number

region::Any

Boundary region number

nspec::Any

Number of species defined in node

coord::Matrix{Tv} where Tv

Grid coordinates

VoronoiFVM.Edge — Type

mutable struct Edge{Tv, Ti} <: VoronoiFVM.AbstractGeometryItem{Tv, Ti}

Structure holding local edge information.

index::Any

Index in grid

node::Vector{Ti} where Ti

Index

region::Any

Inner region number corresponding to edge

nspec::Any

Number of species defined in edge

icell::Any

Number of discretization cell the edge is invoked from

coord::Matrix{Tv} where Tv

Grid coordinates

VoronoiFVM.MatrixUnknowns — Type

struct MatrixUnknowns{T} <: AbstractArray{T, 2}

Wrapper struct for viewing unknowns passed to flux as matrix

u::Vector{T} where T

n1::Int64

VoronoiFVM.Node — Type

mutable struct Node{Tv, Ti} <: VoronoiFVM.AbstractGeometryItem{Tv, Ti}

Structure holding local node information.

index::Any

Index in grid

region::Any

Inner region number

nspec::Any

Number of species defined in node

icell::Any

Number of discretization cell the node is invoked from

coord::Matrix{Tv} where Tv

Grid coordinates

VoronoiFVM.VectorUnknowns — Type

struct VectorUnknowns{T} <: AbstractArray{T, 1}

Wrapper struct for viewing unknowns passed to callback functions

u::Vector{T} where T

n::Int64

offset::Int64

VoronoiFVM.meas — Method

meas(edge::VoronoiFVM.Edge) -> Any

Calculate the length of an edge.

VoronoiFVM.num_species — Method

num_species(edge::VoronoiFVM.Edge) -> Any

Return number of species for edge

VoronoiFVM.unknowns — Method

unknowns(node::VoronoiFVM.BNode, u) -> Any

Construct vector unknowns on bnode

VoronoiFVM.unknowns — Method

unknowns(node::VoronoiFVM.Node, u) -> Any

Construct vector unknowns on node

VoronoiFVM.unknowns — Method

unknowns(edge::VoronoiFVM.Edge, u::Array{T, 1}, i) -> VoronoiFVM.VectorUnknowns{_A} where _A

Construct vector unknowns on edge.

VoronoiFVM.unknowns — Method

unknowns(edge::VoronoiFVM.Edge, u::Array{T, 1}) -> VoronoiFVM.MatrixUnknowns{_A} where _A

Construct matrix of unknowns from edge - these can be used in flux functions with the v0.7.x and v0.8.x syntax to acces data.

VoronoiFVM.viewK — Method

viewK(edge::VoronoiFVM.Edge, u) -> VoronoiFVM.VectorUnknowns{_A} where _A

Solution view on first edge node

VoronoiFVM.viewL — Method

viewL(edge::VoronoiFVM.Edge, u) -> VoronoiFVM.VectorUnknowns{_A} where _A

Solution view on second edge node

Special functions

VoronoiFVM.fbernoulli — Method

fbernoulli(x)

Bernoulli function $B(x)=\frac{x}{e^x-1}$ for exponentially fitted upwinding.

The name fbernoulli has been chosen to avoid confusion with Bernoulli from JuliaStats/Distributions.jl

Returns a real number containing the result.

VoronoiFVM.fbernoulli_pm — Method

fbernoulli_pm(x)

Bernoulli function $B(x)=\frac{x}{e^x-1}$ for exponentially fitted upwind, joint evaluation for positive and negative argument

Usually, we need $B(x), B(-x)$ togehter, and it is cheaper to calculate them together.

Returns two real numbers containing the result for argument x and argument -x.