Physics & special functions

Types and Constants

VoronoiFVM.AbstractData — Type

abstract type AbstractData

Abstract type for user data.

!!! Deprecate There is no benefit in inheriting from AbstractData and this type will be removed in a future release.

VoronoiFVM.AbstractPhysics — Type

abstract type AbstractPhysics

Abstract type for physics.

VoronoiFVM.Physics — Type

struct Physics{Flux<:Function, Reaction<:Function, Storage<:Function, Source<:Function, BReaction<:Function, BStorage<:Function, GenericOperator<:Function, GenericOperatorSparsity<:Function, Data} <: VoronoiFVM.AbstractPhysics

Physics data record.

flux::Function

Flux between neigboring control volumes: flux(f,_u,edge) or flux(f,_u,edge,data) should return in f[i] the flux of species i along the edge joining circumcenters of neigboring control volumes. u=unknowns(_u) returns a 2D array such that for species i, u[i,1] and u[i,2] contain the unknown values at the corresponding ends of the edge.

storage::Function

Storage term (term under time derivative): storage(f,u,node) or storage(f,u,node,data)

It should return in f[i] the storage term for the i-th equation. u[i] contains the value of the i-th unknown.

reaction::Function

Reaction term: reaction(f,u,node) or reaction(f,u,node,data)

It should return in f[i] the reaction term for the i-th equation. u[i] contains the value of the i-th unknown.

source::Function

Source term: source(f,node) or source(f,node,data).

It should return the in f[i] the value of the source term for the i-th equation.

breaction::Function

Boundary reaction term: breaction(f,u,node) or breaction(f,u,node,data) Similar to reaction, but restricted to the inner or outer boundaries.

bstorage::Function

Boundary storage term: bstorage(f,u,node) or bstorage(f,u,node,data) Similar to storage, but restricted to the inner or outer boundaries.

generic_operator::Function

Generic operator generic_operator(f,u,sys). This operator acts on the full solution u of a system. Sparsity is detected automatically unless generic_operator_sparsity is given.

generic_operator_sparsity::Function

Function defining the sparsity structure of the generic operator. This should return the sparsity pattern of the generic_operator.

data::Any

User data (parameters). This allows to pass various parameters to the callback functions.

num_species::Int8

Number of species including boundary species.

VoronoiFVM.Physics — Method

Physics(; num_species, data, flux, reaction, storage, source, breaction, bstorage, generic, generic_sparsity)

Constructor for physics data with default values for the constitutive callback functions.

There are two variants of this constructor. It data is given, all callback functions should accept a last data argument. Otherwise, no data are passed explicitely, and it is assumed that constitutive callbacks take parameters from the closure where the function is defined.

Methods

Base.show — Method

show(io, physics)

Show physics object

VoronoiFVM.hasdata — Method

hasdata(physics)

Check if physics object has data

VoronoiFVM.fbernoulli — Method

fbernoulli(x::Any) -> Any

Bernoulli function implementation for exponentially fitted finite volumes.

The name fbernoulli has been chosen to avoid confusion with Bernoulli from JuliaStats/Distributions.jl

Returns a real number containing the result.

VoronoiFVM.fbernoulli_pm — Method

fbernoulli_pm(x::Any) -> Tuple{Any, Any}

Bernoulli function implementation for exponentially fitted finite volumes, joint evaluation for positive and negative argument

Usually, we need B(x), B(-x) togehter, and it is cheaper to calculate them together.

Returns two real numbers containing the result for argument x and argument -x.