Physics

Types and Constants

VoronoiFVM.AbstractData — Type

abstract type AbstractData

Abstract type for user data.

VoronoiFVM.AbstractPhysics — Type

abstract type AbstractPhysics

Abstract type for physics.

VoronoiFVM.Physics — Type

struct Physics{Flux<:Function, Reaction<:Function, Storage<:Function, Source<:Function, BReaction<:Function, BStorage<:Function, D<:VoronoiFVM.AbstractData} <: VoronoiFVM.AbstractPhysics

Physics data record.

flux::Function

Flux between neigboring control volumes

storage::Function

Storage term (term under time derivative)

reaction::Function

Reaction term

source::Function

" Source term

breaction::Function

Boundary reaction term

bstorage::Function

Boundary storage term

data::VoronoiFVM.AbstractData

User data (parameters)

num_species::Int8

Number of species

VoronoiFVM.Physics — Method

Physics(; num_species, data, flux, reaction, storage, source, breaction, bstorage) -> VoronoiFVM.Physics{typeof(VoronoiFVM.nofunc),typeof(VoronoiFVM.nofunc),typeof(VoronoiFVM.default_storage),typeof(VoronoiFVM.nofunc),typeof(VoronoiFVM.nofunc),typeof(VoronoiFVM.nofunc),VoronoiFVM.NoData}

Constructor for physics data with default values.

Methods

VoronoiFVM.fbernoulli — Method

fbernoulli(x::Real) -> Any

Bernoulli function implementation for exponentially fitted finite volumes.

The name fbernoulli has been chosen to avoid confusion with Bernoulli from JuliaStats/Distributions.jl

Returns a real number containing the result.

VoronoiFVM.fbernoulli_pm — Method

fbernoulli_pm(x::Real) -> Tuple{Any,Any}

Bernoulli function implementation for exponentially fitted finite volumes, joint evaluation for positive and negative argument

Usually, we need B(x), B(-x) togehter, and it is cheaper to calculate them together.

Returns two real numbers containing the result for argument x and argument -x.