Physics & special functions

Physics

VoronoiFVM.AbstractData — Type

abstract type AbstractData

Abstract type for user data.

Compat

Will be removed in future versions

VoronoiFVM.AbstractPhysics — Type

abstract type AbstractPhysics

Abstract type for physics.

VoronoiFVM.Physics — Type

struct Physics

Physics data record with the following fields:

flux::Function

Flux between neigboring control volumes: flux(f,_u,edge) or flux(f,_u,edge,data) should return in f[i] the flux of species i along the edge joining circumcenters of neigboring control volumes. u=unknowns(_u) returns a 2D array such that for species i, u[i,1] and u[i,2] contain the unknown values at the corresponding ends of the edge.

storage::Function

Storage term (term under time derivative): storage(f,u,node) or storage(f,u,node,data)

It should return in f[i] the storage term for the i-th equation. u[i] contains the value of the i-th unknown.

reaction::Function

Reaction term: reaction(f,u,node) or reaction(f,u,node,data)

It should return in f[i] the reaction term for the i-th equation. u[i] contains the value of the i-th unknown.

source::Function

Source term: source(f,node) or source(f,node,data).

It should return the in f[i] the value of the source term for the i-th equation.

bflux::Function

Flux between neighboring control volumes on the boundary

breaction::Function

Boundary reaction term: breaction(f,u,node) or breaction(f,u,node,data) Similar to reaction, but restricted to the inner or outer boundaries.

bsource::Function

Boundary source term: bsource(f,node) or bsource(f,node,data).

It should return in f[i] the value of the source term for the i-th equation.

bstorage::Function

Boundary storage term: bstorage(f,u,node) or bstorage(f,u,node,data) Similar to storage, but restricted to the inner or outer boundaries.

generic_operator::Function

Generic operator generic_operator(f,u,sys). This operator acts on the full solution u of a system. Sparsity is detected automatically unless generic_operator_sparsity is given.

generic_operator_sparsity::Function

Function defining the sparsity structure of the generic operator. This should return the sparsity pattern of the generic_operator.

data::Any

User data (parameters). This allows to pass various parameters to the callback functions.

num_species::Int8

Number of species including boundary species.

VoronoiFVM.Physics — Method

Physics(;num_species=0,
         data=nothing,
         flux,
         reaction,
         storage,
         source,
         breaction,
         bstorage,
         generic,
         generic_sparsity
    )

Constructor for physics data. For the meaning of the optional keyword arguments, see VoronoiFVM.System(grid::ExtendableGrid; kwargs...).

Base.show — Method

show(io, physics)

Show physics object

VoronoiFVM.hasdata — Method

hasdata(physics)

Check if physics object has data

VoronoiFVM.@create_physics_wrappers — Macro

@create_physics_wrappers(physics,node,bnode,edge,bedge)

Create wrapper functions around physics callbacks which fit the API of ForwardDiff.jacobian! and pass the data parameter if necessary. These are meant to be defined before performing assembly loops. The macro creates the follwing variables:

wrapper functions: fluxwrap,storagewrap,reactionwrap,bstoragewrap, breactionwrap

flag variables: issource, isreaction,isbreaction,isbstorage

Edge and node data

VoronoiFVM.AbstractEdge — Type

abstract type AbstractEdge{Tv<:Number, Ti<:Integer} <: VoronoiFVM.AbstractGeometryItem{Tv<:Number, Ti<:Integer}

Abstract type for edges

edge[idim,inode] gives coordinate of node.

VoronoiFVM.AbstractEdgeData — Type

abstract type AbstractEdgeData{T<:Number} <: AbstractArray{T<:Number, 2}

Abstract type for data on edges. u[ispec,inode] accesses value of species at corresponding node.

VoronoiFVM.AbstractGeometryItem — Type

abstract type AbstractGeometryItem{Tv<:Number, Ti<:Integer}

Abstract type for geometry items (node,bnode,edge, bedge)

VoronoiFVM.AbstractNode — Type

abstract type AbstractNode{Tv<:Number, Ti<:Integer} <: VoronoiFVM.AbstractGeometryItem{Tv<:Number, Ti<:Integer}

Abstract type for nodes.

node[idim] gives the the corresponding coordinate.

VoronoiFVM.AbstractNodeData — Type

abstract type AbstractNodeData{T<:Number} <: AbstractArray{T<:Number, 1}

Abstract type for data on nodes. u[ispec] accesses value of species at this node.

VoronoiFVM.BEdge — Type

mutable struct BEdge{Tv, Ti} <: VoronoiFVM.AbstractEdge{Tv, Ti}

Structure holding local edge information.

index::Any

Index in grid

node::Vector

Index

region::Any

Inner region number corresponding to edge

nspec::Any

Number of species defined in edge

icell::Any

Number of discretization cell the edge is invoked from

coord::Matrix

Grid coordinates

bedgenodes::Matrix

bfaceedges::Matrix

bfaceregions::Vector

time::Any

System time

embedparam::Any

Current value of embedding parameter

VoronoiFVM.BNode — Type

mutable struct BNode{Tv, Ti} <: VoronoiFVM.AbstractNode{Tv, Ti}

Structure holding local boundary node information.

index::Any

Index in grid

ibface::Any

BFace number it is called from

ibnode::Any

local node number

region::Any

Boundary region number

cellregions::Vector

nspec::Any

Number of species defined in node

coord::Matrix

Grid coordinates

bfacenodes::Matrix

bfaceregions::Vector

allcellregions::Vector

bfacecells::Adjacency

Dirichlet::Any

time::Any

System time

embedparam::Any

Current value of embedding parameter

dirichlet_value::Vector

VoronoiFVM.Edge — Type

mutable struct Edge{Tv, Ti} <: VoronoiFVM.AbstractEdge{Tv, Ti}

Structure holding local edge information.

index::Any

Index in grid

node::Vector

Index

region::Any

Inner region number corresponding to edge

nspec::Any

Number of species defined in edge

icell::Any

Number of discretization cell the edge is invoked from

coord::Matrix

Grid coordinates

cellx::Matrix

edgenodes::Matrix

cellregions::Vector

has_celledges::Bool

time::Any

System time

embedparam::Any

Current value of embedding parameter

VoronoiFVM.Node — Type

mutable struct Node{Tv, Ti} <: VoronoiFVM.AbstractNode{Tv, Ti}

Structure holding local node information.

index::Any

Index in grid

region::Any

Inner region number

nspec::Any

Number of species defined in node

icell::Any

Number of discretization cell the node is invoked from

coord::Matrix

Grid coordinates

cellnodes::Matrix

Grid cell nodes

cellregions::Vector

Grid cell regions

time::Any

System time

embedparam::Any

Current value of embedding parameter

VoronoiFVM.NodeRHS — Type

struct NodeRHS{T, Tv, Ti} <: VoronoiFVM.AbstractNodeData{T}

RHS data on node.

VoronoiFVM.NodeUnknowns — Type

struct NodeUnknowns{T, Tv, Ti} <: VoronoiFVM.AbstractNodeData{T}

Unknown data on node.

VoronoiFVM.VectorUnknowns — Type

struct VectorUnknowns{T} <: AbstractArray{T, 1}

Wrapper struct for viewing unknowns passed to callback functions

val::Vector

n::Int64

offset::Int64

VoronoiFVM.meas — Method

meas(edge::VoronoiFVM.AbstractEdge) -> Any

Calculate the length of an edge.

VoronoiFVM.num_species — Method

num_species(edge::VoronoiFVM.AbstractEdge) -> Any

Return number of species for edge

VoronoiFVM.unknowns — Method

unknowns(edge::VoronoiFVM.AbstractEdge, u::AbstractArray{T, 1}, i) -> VoronoiFVM.VectorUnknowns

Construct vector unknowns on edge.

VoronoiFVM.viewK — Method

viewK(edge::VoronoiFVM.AbstractEdge, u) -> VoronoiFVM.VectorUnknowns

Solution view on first edge node

VoronoiFVM.viewL — Method

viewL(edge::VoronoiFVM.AbstractEdge, u) -> VoronoiFVM.VectorUnknowns

Solution view on second edge node

Special functions

VoronoiFVM.fbernoulli — Method

fbernoulli(x)

Bernoulli function $B(x)=\frac{x}{e^x-1}$ for exponentially fitted upwinding.

The name fbernoulli has been chosen to avoid confusion with Bernoulli from JuliaStats/Distributions.jl

Returns a real number containing the result.

VoronoiFVM.fbernoulli_pm — Method

fbernoulli_pm(x)

Bernoulli function $B(x)=\frac{x}{e^x-1}$ for exponentially fitted upwind, joint evaluation for positive and negative argument

Usually, we need $B(x), B(-x)$ togehter, and it is cheaper to calculate them together.

Returns two real numbers containing the result for argument x and argument -x.

The error in comparison with the evaluation of the original expression with BigFloat is less than 1.0e-15