Postprocessing

Plotting

Plotting can be performed using the package GridVisualize.jl. This package extends the API with a couple of methods:

GridVisualize.gridplot — Function

gridplot(sys::VoronoiFVM.AbstractSystem; kwargs...) -> Any

Plot grid behind system

GridVisualize.gridplot! — Function

gridplot!(
    vis,
    sys::VoronoiFVM.AbstractSystem;
    kwargs...
) -> Any

Plot grid behind system

GridVisualize.scalarplot — Function

scalarplot(
    sys::VoronoiFVM.AbstractSystem,
    sol::AbstractMatrix;
    species,
    scale,
    kwargs...
) -> Any

Plot one species from solution

scalarplot(
    sys::VoronoiFVM.AbstractSystem,
    sol::VoronoiFVM.AbstractTransientSolution;
    species,
    scale,
    tscale,
    kwargs...
) -> Any

Plot one species from transient solution

GridVisualize.scalarplot! — Function

scalarplot!(
    vis,
    sys::VoronoiFVM.AbstractSystem,
    sol::AbstractMatrix;
    species,
    scale,
    kwargs...
) -> Any

Plot one species from solution

scalarplot!(
    vis,
    sys::VoronoiFVM.AbstractSystem,
    sol::VoronoiFVM.AbstractTransientSolution;
    species,
    scale,
    tscale,
    tlabel,
    kwargs...
) -> Any

Plot one species from transient solution

VoronoiFVM.plothistory — Function

plothistory(tsol; 
            plots=[:timestepsizes,
                   :timestepupdates,
                   :newtonsteps,
                   :newtonupdates], 
            size=(700,600), 
            logmin=1.0e-20,
            Plotter=GridVisualize.default_plotter(),
            kwargs...)

Plot solution history stored in tsol. The plots argument allows to choose which plots are shown.

Grid verification

nondelaunay

Norms & volumes

LinearAlgebra.norm — Function

norm(system, u)
norm(system, u, p)

Calculate Euklidean norm of the degree of freedom vector.

VoronoiFVM.lpnorm — Function

lpnorm(sys, u, p)
lpnorm(sys, u, p, species_weights)

Calculate weighted discrete $L^p$ norm of a solution vector.

VoronoiFVM.l2norm — Function

l2norm(sys, u)
l2norm(sys, u, species_weights)

Calculate weigthed discrete $L^2(\Omega)$ norm of a solution vector.

VoronoiFVM.w1pseminorm — Function

w1pseminorm(sys, u, p)
w1pseminorm(sys, u, p, species_weights)

Calculate weighted discrete $W^{1,p}(\Omega)$ seminorm of a solution vector.

VoronoiFVM.h1seminorm — Function

h1seminorm(sys, u)
h1seminorm(sys, u, species_weights)

Calculate weighted discrete $H^1(\Omega)$ seminorm of a solution vector.

VoronoiFVM.w1pnorm — Function

w1pnorm(sys, u, p)
w1pnorm(sys, u, p, species_weights)

Calculate weighted discrete $W^{1,p}(\Omega)$ norm of a solution vector.

VoronoiFVM.h1norm — Function

h1norm(sys, u)
h1norm(sys, u, species_weights)

Calculate weighted discrete $H^1(\Omega)$ norm of a solution vector.

VoronoiFVM.lpw1pseminorm — Function

lpw1pseminorm(sys, u, p)
lpw1pseminorm(sys, u, p, species_weights)

Calculate weighted discrete $L^p([0,T];W^{1,p}(\Omega))$ seminorm of a transient solution.

VoronoiFVM.l2h1seminorm — Function

l2h1seminorm(sys, u)
l2h1seminorm(sys, u, species_weights)

Calculate weighted discrete $L^2([0,T];H^1(\Omega))$ seminorm of a transient solution.

VoronoiFVM.lpw1pnorm — Function

lpw1pnorm(sys, u, p)
lpw1pnorm(sys, u, p, species_weights)

Calculate weighted discrete $L^p([0,T];W^{1,p}(\Omega))$ norm of a transient solution.

VoronoiFVM.l2h1norm — Function

l2h1norm(sys, u)
l2h1norm(sys, u, species_weights)

Calculate weighted discrete $L^2([0,T];H^1(\Omega))$ norm of a transient solution.

VoronoiFVM.nodevolumes — Function

nodevolumes(system)

Calculate volumes of Voronoi cells.

Solution integrals

VoronoiFVM.integrate — Method

integrate(system,F,U; boundary=false)

Integrate node function (same signature as reaction or storage) F of solution vector region-wise over domain or boundary. The result is an nspec x nregion matrix.

VoronoiFVM.edgeintegrate — Function

edgeintegrate(system,F,U; boundary=false)

Integrate edge function (same signature as flux function) F of solution vector region-wise over domain or boundary. The result is an nspec x nregion matrix.

Nodal flux reconstruction

VoronoiFVM.nodeflux — Function

nodeflux(system, U)

Reconstruction of edge flux as vector function on the nodes of the triangulation. The result can be seen as a piecewiesw linear vector function in the FEM space spanned by the discretization nodes exhibiting the flux density.

The reconstruction is based on the "magic formula" R. Eymard, T. Gallouet, R. Herbin, IMA Journal of Numerical Analysis (2006) 26, 326−353, Lemma 2.4 .

The return value is a dim x nspec x nnodes vector. The flux of species i can e.g. plotted via GridVisualize.vectorplot.

Example:

    ispec=3
    vis=GridVisualizer(Plotter=Plotter)
    scalarplot!(vis,grid,solution[ispec,:],clear=true,colormap=:summer)
    vectorplot!(vis,grid,nf[:,ispec,:],clear=false)
    reveal(vis)

CAVEAT: there is a possible unsolved problem with the values at domain corners in the code. Please see any potential boundary artifacts as a manifestation of this issue and report them.

Boundary flux calculation

VoronoiFVM.TestFunctionFactory — Type

mutable struct TestFunctionFactory

Data structure containing DenseSystem used to calculate test functions for boundary flux calculations.

system::VoronoiFVM.AbstractSystem: Original system

tfsystem::VoronoiFVM.System{Tv, Tc, Ti, Tm, Matrix{Ti}, Matrix{Tv}} where {Tv, Tc, Ti, Tm}: Test function system

control::SolverControl: Solver control

VoronoiFVM.TestFunctionFactory — Method

TestFunctionFactory(
    system::VoronoiFVM.AbstractSystem;
    control
) -> TestFunctionFactory

Constructor for TestFunctionFactory from System

VoronoiFVM.integrate — Method

integrate(system, tf, U; kwargs...)

Calculate test function integral for steady state solution.

VoronoiFVM.integrate — Method

integrate(system, tf, U, Uold, tstep; params)

Calculate test function integral for transient solution.

VoronoiFVM.integrate_stdy — Method

integrate_stdy(system, tf, U)

Steady state part of test function integral.

VoronoiFVM.integrate_tran — Method

integrate_tran(system, tf, U)

Calculate transient part of test function integral.

VoronoiFVM.testfunction — Method

testfunction(factory::TestFunctionFactory, bc0, bc1) -> Any

Create testfunction which has Dirichlet zero boundary conditions for boundary regions in bc0 and Dirichlet one boundary conditions for boundary regions in bc1.

Impedance calculatiom

Impedance calculation can be seen as a postprocessing step after the solution of the unexcited stationary system.

VoronoiFVM.AbstractImpedanceSystem — Type

abstract type AbstractImpedanceSystem{Tv<:Number}

Abstract type for impedance system.

VoronoiFVM.ImpedanceSystem — Type

mutable struct ImpedanceSystem{Tv} <: VoronoiFVM.AbstractImpedanceSystem{Tv}

Concrete type for impedance system.

sysnzval::AbstractArray{Complex{Tv}, 1} where Tv: Nonzero pattern of time domain system matrix

storderiv::AbstractMatrix: Derivative of storage term

matrix::AbstractArray{Complex{Tv}, 2} where Tv: Complex matrix of impedance system

F::AbstractArray{Complex{Tv}, 2} where Tv: Right hand side of impedance system

U0::AbstractMatrix: Stationary state

VoronoiFVM.ImpedanceSystem — Method

ImpedanceSystem(system, U0, excited_spec, excited_bc)

Construct impedance system from time domain system sys and steady state solution U0 under the assumption of a periodic perturbation of species excited_spec at boundary excited_bc.

VoronoiFVM.ImpedanceSystem — Method

ImpedanceSystem(system, U0; λ0)

Construct impedance system from time domain system sys and steady state solution U0

CommonSolve.solve! — Method

solve!(UZ, impedance_system, ω)

Solve the impedance system for given frequency ω.

VoronoiFVM.freqdomain_impedance — Method

freqdomain_impedance(
    impedance_system,
    ω,
    U0,
    excited_spec,
    excited_bc,
    excited_bcval,
    dmeas_stdy,
    dmeas_tran
)

Calculate reciprocal value of impedance.

excitedspec,excitedbc,excited_bcval are ignored.

Warning

This is deprecated: use impedance.

VoronoiFVM.impedance — Method

impedance(impedance_system,ω, U0 ,
          excited_spec, excited_bc, excited_bcval,
           dmeas_stdy,
           dmeas_tran 
           )
    

Calculate impedance.

ω: frequency

U0: steady state slution

dmeas_stdy: Derivative of steady state part of measurement functional

dmeas_tran Derivative of transient part of the measurement functional

VoronoiFVM.measurement_derivative — Method

measurement_derivative(system, measurement_functional, U0)

Calculate the derivative of the scalar measurement functional at steady state U0

Usually, this functional is a test function integral. Initially, we assume that its value depends on all unknowns of the system.

VoronoiFVM.unknowns — Method

unknowns(impedance_system)

Create a vector of unknowns of the impedance system