Postprocessing

Plotting

Plotting can be performed using the package GridVisualize.jl.

Solution integrals

Boundary flux calculation

VoronoiFVM.TestFunctionFactory — Type

mutable struct TestFunctionFactory{Tv}

Data structure containing DenseSystem used to calculate test functions for boundary flux calculations.

system::VoronoiFVM.AbstractSystem{Tv} where Tv

Original system

tfsystem::VoronoiFVM.System{Tv, Ti, Tm, Matrix{Ti}, Matrix{Tv}} where {Tv, Ti, Tm}

Test function system

VoronoiFVM.TestFunctionFactory — Method

Constructor for TestFunctionFactory from System

VoronoiFVM.integrate — Method

integrate(system, tf, U, Uold, tstep)

Calculate test function integral for transient solution.

VoronoiFVM.integrate — Method

integrate(system, tf, U)

Calculate test function integral for steady state solution.

VoronoiFVM.integrate_stdy — Method

Steady state part of test function integral.

VoronoiFVM.integrate_tran — Method

integrate_tran(system, tf, U)

Calculate transient part of test function integral.

VoronoiFVM.testfunction — Method

testfunction(factory::VoronoiFVM.TestFunctionFactory{Tv}, bc0, bc1) -> Any

Create testfunction which has Dirichlet zero boundary conditions for boundary regions in bc0 and Dirichlet one boundary conditions for boundary regions in bc1.

Impedance calculatiom

Impedance calculation can be seen as a postprocessing step after the solution of the unexcited stationary system.

VoronoiFVM.AbstractImpedanceSystem — Type

abstract type AbstractImpedanceSystem{Tv<:Number}

Abstract type for impedance system.

VoronoiFVM.ImpedanceSystem — Type

mutable struct ImpedanceSystem{Tv} <: VoronoiFVM.AbstractImpedanceSystem{Tv}

Concrete type for impedance system.

sysnzval::AbstractArray{Complex{Tv}, 1} where Tv

Nonzero pattern of time domain system matrix

grid::Any

Discretization grid

storderiv::AbstractMatrix

Derivative of storage term

matrix::AbstractArray{Complex{Tv}, 2} where Tv

Complex matrix of impedance system

F::AbstractArray{Complex{Tv}, 2} where Tv

Right hand side of impedance system

U0::AbstractMatrix

Stationary state

VoronoiFVM.ImpedanceSystem — Method

ImpedanceSystem(system, U0, excited_spec, excited_bc)

Construct impedance system from time domain system sys and steady state solution U0 under the assumption of a periodic perturbation of species excited_spec at boundary excited_bc.

VoronoiFVM.impedance — Method

impedance(impedance_system,ω, U0 ,
          excited_spec, excited_bc, excited_bcval,
           dmeas_stdy,
           dmeas_tran 
           )
    

Calculate impedance.

ω: frequency

U0: steady state slution

dmeas_stdy: Derivative of steady state part of measurement functional

dmeas_tran Derivative of transient part of the measurement functional

VoronoiFVM.measurement_derivative — Method

measurement_derivative(system, measurement_functional, U0)

Calculate the derivative of the scalar measurement functional at steady state U0

Usually, this functional is a test function integral. Initially, we assume that its value depends on all unknowns of the system.

VoronoiFVM.solve! — Method

solve!(UZ, impedance_system, ω)

Solve the impedance system for given frequency ω.

VoronoiFVM.unknowns — Method

Create a vector of unknowns of the impedance system