Solution objects

Dense solution arrays

VoronoiFVM.DenseSolutionArray — Type

const DenseSolutionArray=Matrix

Dense storage of solution

VoronoiFVM._add — Method

_add(U::Array{Tv, 2}, idof, val) -> Any

Add residual value into global degree of freedom

VoronoiFVM.dof — Method

dof(a, ispec, K)

Get degree of freedom number

VoronoiFVM.values — Method

values(a)

Array of values in solution array.

Sparse solution arrays

VoronoiFVM.SparseSolutionArray — Type

struct SparseSolutionArray{Tv, Ti} <: AbstractArray{Tv, 2}

Struct holding solution information for SparseSystem. Solution is stored in a sparse matrix structure.

This class plays well with the abstract array interface.

Fields:

node_dof::SparseArrays.SparseMatrixCSC{Tv, Ti} where {Tv, Ti}: Sparse matrix holding actual data.

Base.copy — Method

copy(this)

Create a copy of sparse solution array

Base.getindex — Method

getindex(a, ispec, inode)

Accessor for sparse solution array.

Base.setindex! — Method

setindex!(a, v, ispec, inode)

Accessor for sparse solution array.

Base.similar — Method

similar(this)

Create a similar unintialized sparse solution array

Base.size — Method

size(a)

Return size of sparse solution array.

VoronoiFVM._add — Method

_add(U::VoronoiFVM.SparseSolutionArray, idof, val) -> Any

Add residual value into global degree of freedom

VoronoiFVM.dof — Method

dof(a, i, j)

Get number of degree of freedom. Return 0 if species is not defined in node.

VoronoiFVM.getdof — Method

getdof(a, i)

Return value for degree of freedom.

VoronoiFVM.setdof! — Method

setdof!(a, v, i)

Set value for degree of freedom.

VoronoiFVM.values — Method

values(a)

Array of values in sparse solution array.

Transient solution

VoronoiFVM.AbstractTransientSolution — Type

abstract type AbstractTransientSolution{T, N, A, B} <: RecursiveArrayTools.AbstractDiffEqArray{T, N, A}

Abstract type for transient solution

VoronoiFVM.TransientSolution — Type

mutable struct TransientSolution{T, N, A, B} <: VoronoiFVM.AbstractTransientSolution{T, N, A, B}

Transient solution structure

Fields

u::Any: Vector of solutions

t::Any: Vector of times

Interface

Object of this type adhere to the AbstractDiffEqArray interface. For indexing and interpolation, see https://diffeq.sciml.ai/stable/basics/solution/.

In particular, a TransientSolution sol can be accessed as follows:

sol[i] contains the solution for timestep i

sol[ispec,:,i] contains the solution for component ispec at timestep i

sol(t) returns a (linearly) interpolated solution value for t.

sol.t[i] is the corresponding time for timestep i

sol[ispec,ix,i] refers to solution of component ispec at node ix at moment i

VoronoiFVM.TransientSolution — Method

TransientSolution(t0,inival;
                  in_memory=true,
                  keep_open=true,
                  fname=tempname(pwd())*".jld2"

Constructor of transient solution with initial value and inital time.

in_memory: if true (default), data are kept in main memory, otherwise on disk (via JLD2)

keep_open: if true, disk file is not closed during the existence of the object

fname: file name for the disk file

VoronoiFVM.VectorOfDiskArrays — Method

VectorOfDiskArrays(firstobj:AbstractArray;
                   keep_open=true,
                   fname= fname=tempname(pwd())*".jld2")

Constructor of vector of arrays stored on disk (via JLD2).

keep_open: if true, disk file is not closed during the existence of the object

fname: file name for the disk file

The disk file is automatically removed if the object is garbage collected.