Solution methods

Stationary and transient solvers

VoronoiFVM.embed! — Method

function embed!(solution, inival, system; 
                control=NewtonControl(),
                pre=function(sol,p) end,
                post=function(sol,p) end
)

Solve stationary problem via parameter embedding, calling solve! for increasing values of the parameter p from interval (0,1). The user is responsible for the interpretation of the parameter.

The optional pre() callback can be used to communicate its value to the system.

The optional post() callback method can be used to perform various postprocessing steps.

If control.handle_error is true, if solve! throws an error, stepsize control.Δp is lowered, and solve! is called again with a smaller parameter value. If control.Δp<control.Δp_min, embed! is aborted with error.

VoronoiFVM.evolve! — Method

function evolve!(solution, inival, system, times;
                 control=NewtonControl(), 
                 pre=function(sol,t) end,   
                 post=function(sol,oldsol, t, Δt) end,
                 sample=function(sol,t) end,
                 delta=(u,v,t, Δt)->norm(sys,u-v,Inf)
)

Use implicit Euler method + damped Newton's method to solve time dependent problem. Time step control is performed according to the data in control. All times in times are reached exactly.

Callbacks:

pre is invoked before each time step

post is invoked after each time step

sample is called for all times in times[2:end].

delta is used to control the time step.

VoronoiFVM.solve! — Method

solve!(solution, inival, system; 
    control=NewtonControl(), 
    tstep=Inf, 
    log=false)

Perform solution of stationary problem(if tstep==Inf) or of one step of the implicit Euler method using Newton's method with inival as initial value. The method writes into solution.

It returns solution or, if log==true, a tuple of solution and a vector containing the residual history of Newton's method.

VoronoiFVM.solve — Method

function solve(inival, system, times;
               control=NewtonControl(), 
               pre=function(sol,t) end,   
               post=function(sol,oldsol, t, Δt) end,
               sample=function(sol,t) end,
               delta=(u,v,t, Δt)->norm(sys,u-v,Inf),
               store_all=true,
               in_memory=true
)

Use implicit Euler method + damped Newton's method to solve time dependent problem. Time step control is performed according to the data in control. All times in the vector times are reached exactly.

Callbacks:

pre is invoked before each time step

post is invoked after each time step

sample is called for all times in times[2:end].

delta is used to control the time step.

If store_all==true, all timestep solutions are stored. Otherwise, only solutions for the moments defined in times are stored.

Returns a transient solution object sol containing the stored solution, see TransientSolution

VoronoiFVM.solve — Method

solve(inival, system; 
      control=NewtonControl(), 
      tstep=Inf, 
      log=false)

Perform solution of stationary problem(if tstep==Inf) or one step of the implicit Euler method using Newton's method with inival as initial value. It returns a solution array or, if log==true, a tuple of solution and a vector containing the residual history of Newton's method.

VoronoiFVM.value — Function

Extract value from dual number. Use to debug physics callbacks. Re-exported from ForwardDiff.jl

Solver control

VoronoiFVM.NewtonControl — Type

mutable struct NewtonControl

Control parameters for Newton method.

Newton's method solves $F(u)=0$ by the iterative procedure $u_{i+1}=u_{i} - d_i F'(u_i)^{-1}F(u_i)$ starting with some inital value $u_0$, where $d_i$ is the damping.

tol_absolute::Float64

Tolerance (in terms of norm of Newton update): terminate if $\Delta_i=||u_{i+1}-u_i||_\infty <$ tol_absolute.

Default value: 1.0e-10.

tol_relative::Float64

Tolerance (relative to the first residual): terminate if $\Delta_i/\Delta_0<$ tol_relative.

Default value: 1.0e-10.

tol_round::Float64

Tolerance for roundoff error detection: terminate if $|\;||u_{i+1}||_1 - ||u_{i}||_1\;|/ ||u_{i}||_1<$ tol_round occured max_round times in a row.

Default value: 1.0e-10.

tol_mono::Float64

Tolerance for monotonicity test: terminat with error if $\Delta_i/\Delta_{i-1}>$ 1/tol_mono.

Default value: 1.0e-3

damp_initial::Float64

Initial damping parameter $d_0$.

Default value: 1.0.

To handle convergence problems, set this to a value less than 1.

damp_growth::Float64

Damping parameter growth factor: $d_{i+1}=\min(d_i\cdot$ max_growth $,1)$

Default value: 1.2

Generally it should be set to a value between 1 and 2.

max_iterations::Int32

Maximum number of iterations.

Default value: 100

max_lureuse::Int32

Maximum number of reuses of lu factorization. It this value is 0, linear systems are solved by a sparse direct solver, and it's LU factorization is called in every Newton step.

Otherwise, a BICGstab iterative method is used for linear system solution with an LU factorization as preconditioner which is updated only every max_lureuse Newton step.

Default value: 0.

max_round::Int32

Maximum number of consecutive iterations within roundoff error tolerance

Default value: 1000 (effectively switching of this criterion).

tol_linear::Float64

Tolerance of iterative linear solver.

Default value: 1.0e-4.

factorization::ExtendableSparse.AbstractFactorization

Factorization kind for linear sytems (see ExtendableSparse.jl)

verbose::Bool

Verbosity flag

Default value: false.

handle_exceptions::Bool

Handle exceptions in embed! and evolve! methods. If true, exceptions in Newton solves are catched, the embedding resp. time step is lowered, and solution is retried.

Default value: false

Δp::Float64

Initial parameter step for embed! method.

Default value: 1.0

Δp_max::Float64

Maximal parameter step for embed! method

Default value: 1.0

Δp_min::Float64

Minimal parameter step for embed! method.

Default value: 1.0e-3

Δt::Float64

Time step for evolve! method.

Default value: 0.1

Δt_max::Float64

Maximal time step for evolve! method.

Default value: 1

Δt_min::Float64

Minimal time step for evolve! method.

Default value: 1.0e-3

Δt_grow::Float64

Maximal step size growth for evolve! method.

Default: 1.2

Δu_opt::Float64

Optimal size of update for evolve! method. The algorithm keeps this value approximately constant.

Default: 0.1

force_first_step::Bool

force first timestep

edge_cutoff::Float64

Edge cutoff for rectangular triangles.

Default value: 0.0.

umfpack_pivot_tolerance::Float64

Pivot tolerance for umfpack

Default value: 0.1.

VoronoiFVM.NewtonControl — Method

NewtonControl(this) -> Any

Default constructor

VoronoiFVM.fixed_timesteps! — Method

timesteps!(control,Δt; grow=1.0)

Modify control data such that the time steps are fixed to a geometric sequence such that Δtnew=Δtold*grow

Interface to DifferentialEquations.jl

VoronoiFVM.eval_jacobian! — Method

eval_jacobian!(J, u, sys, t)

Interpret the discrete problem as an ODE/DAE problem. Provide the jacobi matrix calculation function for DifferentialEquations.jl.

VoronoiFVM.eval_rhs! — Method

eval_rhs!(du, u, sys, t)

Interpret the discrete problem as an ODE/DAE problem. Provide the rhs function for DifferentialEquations.jl.

VoronoiFVM.jac_prototype — Method

jac_prototype(sys)

Complete the system and provide the jacobi matrix as prototype for the jacobian for use with DifferentialEquations.jl.

VoronoiFVM.mass_matrix — Method

Calculate the mass matrix for use with DifferentialEquations.jl.

VoronoiFVM.solve — Method

solve(DifferentialEquations, inival, system, tspan; 
      solver=nothing, 
      kwargs...)

Use a timestepping scheme from DifferentialEquations.jl to perform transient solution of the system.

The system must have a constant diagonal mass matrix.

Any stiff solver capable to work with mass matrices is possible. If the system contains elliptic equations, it needs to be a DAE solver able to work with mass matrices. See a list here.

The constructed solver should be passed as solver= kwarg. By default (if nothing is passed), the Rosenbrock23 solver is used.

All other kwargs are passed to DifferentialEquations.solve(), see here for an overview.

In order to avoid a package dependency on DifferentialEquations, the module corresponding to that package is passed as the first argument.

Returns a transient solution object sol containing stored solutions, see TransientSolution.