System

The computational grid is required is assumed to correspond to a domain $\Omega=\cup_{r=1}^{n_\Omega} \Omega_r$

with boundary $\partial\Omega=\Gamma=\cup_{b=1}^{n_\Gamma} \Gamma_b$.

The subdomains $\Omega_r$ are called "regions" and the boundary subdomains $\Gamma_b$ are called "boundary regions".

On this complex of domains "lives" a number of species which are either attached to a number of regions or to a number of boundary regions.

Grids for VoronoiFVM are managed by the package ExtendableGrids.jl.

VoronoiFVM.AbstractSystem — Type

abstract type AbstractSystem{Tv<:Number, Ti<:Integer, Tm<:Integer}

Abstract type for finite volume system structure.

VoronoiFVM.DenseSystem — Type

const DenseSystem

Type alias for system with dense matrix based species management

VoronoiFVM.DenseSystem — Method

DenseSystem(grid, physics; matrixindextype)

Constructor for DenseSystem.

VoronoiFVM.SparseSystem — Type

const SparseSystem

Type alias for system with sparse matrix based species management

VoronoiFVM.SparseSystem — Method

SparseSystem(grid, physics; matrixindextype)

Constructor for SparseSystem.

VoronoiFVM.System — Type

mutable struct System{Tv, Ti, Tm, TSpecMat<:(AbstractMatrix{T} where T), TSolArray<:(AbstractMatrix{T} where T)} <: VoronoiFVM.AbstractSystem{Tv, Ti, Tm}

Structure holding data for finite volume system solution.

Information on species distribution is kept in sparse or dense matrices matrices and, correspondingly, the solution array is of type SparseSolutionArray or matrix, respectively.

In the case of sparse unknown storage, the system matrix handles exactly those degrees of freedom which correspond to unknowns. However, handling of the sparse matrix structures for the bookkeeping of the unknowns creates overhead.

grid::ExtendableGrids.ExtendableGrid{Tv, Ti} where {Tv, Ti}

Grid

physics::VoronoiFVM.Physics

Physics data

boundary_values::Matrix{Tv} where Tv

Array of boundary values

boundary_factors::Matrix{Tv} where Tv

Array of boundary factors

region_species::AbstractMatrix{T} where T

Matrix containing species numbers for inner regions

bregion_species::AbstractMatrix{T} where T

Matrix containing species numbers for boundary regions

node_dof::AbstractMatrix{T} where T

Matrix containing degree of freedom numbers for each node

matrix::ExtendableSparse.ExtendableSparseMatrix{Tv, Tm} where {Tv, Tm}

Jacobi matrix for nonlinear problem

factorization::Union{Nothing, ExtendableSparse.AbstractFactorization{Tv, Tm}} where {Tv, Tm}

Matrix factorization

species_homogeneous::Bool

Flag which says if the number of unknowns per node is constant

num_quantities::Any

Number of quantities defined on system

update::AbstractMatrix{T} where T

Solution vector holding Newton update

residual::AbstractMatrix{T} where T

Solution vector holding Newton residual

cellnodefactors::Matrix{Tv} where Tv

Precomputed geometry factors for cell nodes

celledgefactors::Matrix{Tv} where Tv

Precomputed geometry factors for cell edges

bfacenodefactors::Matrix{Tv} where Tv

Precomputed geometry factors for boundary nodes

bfaceedgefactors::Matrix{Tv} where Tv

Precomputed geometry factors for boundary edges

generic_matrix::SparseArrays.SparseMatrixCSC

Sparse matrix for generic operator handling

generic_matrix_colors::Vector{T} where T

Sparse matrix colors for generic operator handling

uhash::UInt64

Hash value of latest unknowns vector the assembly was called with

allocs::Int64

Data for allocation check

VoronoiFVM.System — Method

function System(grid,physics;
                unknown_storage=:dense;
                matrixindextype=Int32)

Create Finite Volume System.

grid: 1D/2D/3D discretization grid

physics: Physics struct containing node and edge callbacks

unknown_storage: string or symbol:

:dense : solution vector is an nspecies x nnodes dense matrix

:sparse : solution vector is an nspecies x nnodes sparse matrix

matrixindextype : Index type for sparse matrices created in the system

Base.reshape — Method

reshape(v, system)

Reshape vector to fit as solution to system.

LinearAlgebra.norm — Method

norm(system, u, p)

Calculate norm, paying attention to species distribution over regions

VoronoiFVM.addzrows — Method

addzrows(matrix,maxrow)

Return matrix with number of rows increased to maxrow, and set the new elements to zero.

VoronoiFVM.boundary_dirichlet! — Method

boundary_dirichlet!(system, ispec, ibc, v)

Set Dirichlet boundary condition for species ispec at boundary ibc:

$u_{ispec}=v$ on $\Gamma_{ibc}$

VoronoiFVM.boundary_neumann! — Method

boundary_neumann!(system, ispec, ibc, v)

Set Neumann boundary condition for species ispec at boundary ibc:

$\mathrm{flux}_{ispec}\cdot \vec n=v$ on $\Gamma_{ibc}$

VoronoiFVM.boundary_robin! — Method

boundary_robin!(system, ispec, ibc, α, v)

Set Robin boundary condition for species ispec at boundary ibc:

$\mathrm{flux}_{ispec}\cdot \vec n + \alpha u_{ispec}=v$ on $\Gamma_{ibc}$

VoronoiFVM.check_allocs! — Method

check_allocs!(system,true_or_false)

Enable/disable checking for time-consuming allocations in the assembly loop. By default, this check is switched off. By setting the environment variable ENV["VORONOIFVM_CHECK_ALLOCS"]="true", this default can be changed.

Unless the matrix pattern changes, there shouldn't occur any allocations in this loop. The check method is aware of matrix pattern changes. As a consequence, allocations in the assembly loop are mostly due to type instabilities in physics callbacks, see the the discussion here. Type instabilities can be debugged via the @time macro applied to expressions in a physics callback.

The following cases provide some ideas where to look for reasons of the problem and possible remedies:

Case 1: a parameter changes its value, and Julia is not sure about the type.

eps=1.0

flux(f,_u,edge)
    u=unkowns(edge,_u)
    f[1]=eps*(u[1,1]-[1,2])
end
... solve etc ...
eps=2.0

Remedy: use a type annotation eps::Float64=... to signalize your intent to Julia. This behaviour is explained in the Julia documentation.

Case 2: variables in the closure have the same name as a variable introduced in a callback.

flux(f,_u,edge)
    u=unkowns(edge,_u)
    f[1]=(u[1,1]-[1,2])
end

... create etc ...

u=solve(...)

Remedy: rename e.g. u=solve() to sol=solve()

VoronoiFVM.data — Method

data(system)

Retrieve user data record.

VoronoiFVM.enable_boundary_species! — Method

enable_boundary_species!(system,ispec,regions)

Add species ispec to a list of boundary regions. Species numbers for bulk and boundary species have to be distinct.

VoronoiFVM.enable_species! — Method

enable_species!(system,ispec,regions)

Add species ispec to a list of bulk regions. Species numbers for bulk and boundary species have to be distinct.

VoronoiFVM.generic_operator_sparsity! — Method

generic_operator_sparsity!(system, sparsematrix)

Set generic operator sparsity, in the case where a generic operator has been defined in physics.

VoronoiFVM.increase_num_species! — Method

 increase_num_species!(system,maxspec)

Increase number of species in system to maxspec by adding new rows to all relevant matrices.

VoronoiFVM.is_boundary_species — Method

is_boundary_species(AbstractSystem, ispec) -> Bool

Check if species number corresponds to a boundary species.

VoronoiFVM.is_bulk_species — Method

is_bulk_species(AbstractSystem, ispec) -> Bool

Check if species number corresponds to a bulk species.

VoronoiFVM.isdof — Method

isdof(system, ispec, inode)

Check if degree of freedom is defined.

VoronoiFVM.num_dof — Method

num_dof(system)

Number of degrees of freedom for system.

VoronoiFVM.num_species — Method

num_species(a)

Number of species (size of first dimension) of solution array.

VoronoiFVM.num_species — Method

num_species(system)

Number of species in system

VoronoiFVM.unknowns — Method

unknowns(Tu, sys; inival)

Create a solution vector for system with elements of type Tu. If inival is not specified, the entries of the returned vector are undefined.

VoronoiFVM.unknowns — Method

unknowns(system; inival)

Create a solution vector for system. If inival is not specified, the entries of the returned vector are undefined.

VoronoiFVM.update_grid! — Method

update_grid!(system; grid=system.grid)

Update grid (e.g. after rescaling of coordinates).