System

Types and Constants

VoronoiFVM.AbstractSystem — Type

abstract type AbstractSystem{Tv<:Number, Ti<:Integer, Tm<:Integer}

Abstract type for finite volume system structure

VoronoiFVM.DenseSolutionArray — Type

const DenseSolutionArray=Matrix

Dense storage of solution

VoronoiFVM.SparseSolutionArray — Type

struct SparseSolutionArray{Tv, Ti} <: AbstractArray{Tv, 2}

Struct holding solution information for SparseSystem. Solution is stored in a sparse matrix structure.

This class plays well with the abstract array interface.

Fields:

node_dof::SparseArrays.SparseMatrixCSC{Tv, Ti} where {Tv, Ti}

Sparse matrix holding actual data.

VoronoiFVM.AbstractTransientSolution — Type

abstract type AbstractTransientSolution{T, N, A, B} <: RecursiveArrayTools.AbstractDiffEqArray{T, N, A}

Abstract type for transient solution

VoronoiFVM.TransientSolution — Type

mutable struct TransientSolution{T, N, A, B} <: VoronoiFVM.AbstractTransientSolution{T, N, A, B}

Transient solution structure

Fields

u::Any

Vector of solutions

t::Any

Vector of times

Interface

Object of this type adhere to the AbstractDiffEqArray interface. For indexing and interpolation, see https://diffeq.sciml.ai/stable/basics/solution/.

In particular, a TransientSolution sol can be accessed as follows:

sol[it] contains the solution for timestep i

sol[ispec,:,it] contains the solution for component ispec at timestep i

sol(t) returns a (linearly) interpolated solution value for t.

sol.t[it] is the corresponding time

sol[ispec,ix,it] refers to solution of component ispec at node ix at moment it

VoronoiFVM.TransientSolution — Method

TransientSolution(t0,inival;
                  in_memory=true,
                  keep_open=true,
                  fname=tempname(pwd())*".jld2"

Constructor of transient solution with initial value and inital time.

in_memory: if true (default), data are kept in main memory, otherwise on disk (via JLD2)

keep_open: if true, disk file is not closed during the existence of the object

fname: file name for the disk file

VoronoiFVM.VectorOfDiskArrays — Method

VectorOfDiskArrays(firstobj:AbstractArray;
                   keep_open=true,
                   fname= fname=tempname(pwd())*".jld2")

Constructor of vector of arrays stored on disk (via JLD2).

keep_open: if true, disk file is not closed during the existence of the object

fname: file name for the disk file

The disk file is automatically removed if the object is garbage collected.

VoronoiFVM.DenseSystem — Type

mutable struct DenseSystem{Tv, Ti, Tm} <: VoronoiFVM.AbstractSystem{Tv, Ti, Tm}

Structure holding data for finite volume system solution. Information on species distribution is kept in dense matrices, and the solution array is of type Array{2}.

Unlike in the SparseSystem, the system matrix handles exactly those degrees of freedom which correspond to unknowns, and dummy degrees of freedom where unknowns are not defined. Handling of the sparse matrix structures for the bookeeping of the unknowns has less overhead, but additional dummy equations are added to the system matrix.

grid::Any

Grid

physics::VoronoiFVM.Physics

Physics data

boundary_values::Matrix{Tv} where Tv

Array of boundary values

boundary_factors::Matrix{Tv} where Tv

Array of boundary factors

region_species::Matrix{Int8}

Full matrix containing species numbers for inner regions

bregion_species::Matrix{Int8}

Full matrix containing species numbers for boundary regions

node_dof::Matrix{Int8}

Full matrix containing degree of freedom numbers for each node

matrix::ExtendableSparse.ExtendableSparseMatrix{Tv, Tm} where {Tv, Tm}

Jacobi matrix for nonlinear problem

species_homogeneous::Bool

Flag which says if the number of unknowns per node is constant

update::Matrix{Tv} where Tv

Solution vector holding Newton update

residual::Matrix{Tv} where Tv

Solution vector holding Newton residual

cellnodefactors::Matrix{Tv} where Tv

Precomputed geometry factors for cell nodes

celledgefactors::Matrix{Tv} where Tv

Precomputed geometry factors for cell edges

bfacenodefactors::Matrix{Tv} where Tv

Precomputed geometry factors for boundary nodes

generic_matrix::SparseArrays.SparseMatrixCSC

Sparse matrix for generic operator handling

generic_matrix_colors::Vector{T} where T

Sparse matrix colors for generic operator handling

uhash::UInt64

Hash value of latest unknowns vector the assembly was called with

VoronoiFVM.DenseSystem — Method

DenseSystem(grid::Any, physics::VoronoiFVM.Physics; matrixindextype) -> VoronoiFVM.DenseSystem{_A, _B, Int32} where {_A, _B}

Constructor for DenseSystem.

VoronoiFVM.SparseSystem — Type

mutable struct SparseSystem{Tv, Ti, Tm} <: VoronoiFVM.AbstractSystem{Tv, Ti, Tm}

Structure holding data for finite volume system solution.

Information on species distribution is kept in sparse matrices, and the solution array is of type SparseSolutionArray, i.e. effectively it is a sparse matrix.

Unlike in the DenseSystem, the system matrix handles exactly those degrees of freedom which correspond to unknowns. However, handling of the sparse matrix structures for the bookkeeping of the unknowns creates overhead.

grid::Any

Grid

physics::VoronoiFVM.Physics

Physics data

boundary_values::Matrix{Tv} where Tv

Array of boundary values

boundary_factors::Matrix{Tv} where Tv

Array of boundary factors

region_species::SparseArrays.SparseMatrixCSC{Int8, Ti} where Ti

Sparse matrix containing species numbers for inner regions

bregion_species::SparseArrays.SparseMatrixCSC{Int8, Ti} where Ti

Sparse matrix containing species numbers for boundary regions

node_dof::SparseArrays.SparseMatrixCSC{Int8, Tm} where Tm

Sparse matrix containing degree of freedom numbers for each node

matrix::ExtendableSparse.ExtendableSparseMatrix{Tv, Tm} where {Tv, Tm}

Jacobi matrix for nonlinear problem

species_homogeneous::Bool

Flag which says if the number of unknowns per node is constant

update::VoronoiFVM.SparseSolutionArray{Tv, Tm} where {Tv, Tm}

Solution vector holding Newton update

residual::VoronoiFVM.SparseSolutionArray{Tv, Tm} where {Tv, Tm}

Solution vector holding Newton residual

cellnodefactors::Matrix{Tv} where Tv

Precomputed geometry factors for cell nodes

celledgefactors::Matrix{Tv} where Tv

Precomputed geometry factors for cell edges

bfacenodefactors::Matrix{Tv} where Tv

Precomputed geometry factors for boundary nodes

generic_matrix::SparseArrays.SparseMatrixCSC

Sparse matrix for generic operator handling

generic_matrix_colors::Vector{T} where T

Sparse matrix colors for generic operator handling

uhash::UInt64

Hash value of latest unknowns vector the assembly was called with

VoronoiFVM.SparseSystem — Method

SparseSystem(grid::Any, physics::VoronoiFVM.Physics; matrixindextype) -> VoronoiFVM.SparseSystem{_A, _B, Int32} where {_A, _B}

Constructor for SparseSystem.

VoronoiFVM.BNode — Type

mutable struct BNode{Tv, Ti} <: VoronoiFVM.AbstractGeometryItem{Tv, Ti}

Structure holding local boundary node information.

index::Any

Index in grid

region::Any

Boundary region number

nspec::Any

Number of species defined in node

coord::Matrix{Tv} where Tv

Grid coordinates

VoronoiFVM.Edge — Type

mutable struct Edge{Tv, Ti} <: VoronoiFVM.AbstractGeometryItem{Tv, Ti}

Structure holding local edge information.

index::Any

Index in grid

node::Vector{Ti} where Ti

Index

region::Any

Inner region number corresponding to edge

nspec::Any

Number of species defined in edge

icell::Any

Number of discretization cell the edge is invoked from

coord::Matrix{Tv} where Tv

Grid coordinates

VoronoiFVM.MatrixUnknowns — Type

struct MatrixUnknowns{T} <: AbstractArray{T, 2}

Wrapper struct for viewing unknowns passed to flux as matrix

u::Vector{T} where T

n1::Int64

VoronoiFVM.Node — Type

mutable struct Node{Tv, Ti} <: VoronoiFVM.AbstractGeometryItem{Tv, Ti}

Structure holding local node information.

index::Any

Index in grid

region::Any

Inner region number

nspec::Any

Number of species defined in node

icell::Any

Number of discretization cell the node is invoked from

coord::Matrix{Tv} where Tv

Grid coordinates

VoronoiFVM.VectorUnknowns — Type

struct VectorUnknowns{T} <: AbstractArray{T, 1}

Wrapper struct for viewing unknowns passed to callback functions

u::Vector{T} where T

n::Int64

offset::Int64

VoronoiFVM.AssemblyError — Type

struct AssemblyError <: Exception

Exception thrown if error occured during assembly (e.g. domain error)

VoronoiFVM.ConvergenceError — Type

struct ConvergenceError <: Exception

Exception thrown if Newton's method convergence fails.

VoronoiFVM.EmbeddingError — Type

struct EmbeddingError <: Exception

Exception thrown if embedding fails

VoronoiFVM.FactorizationError — Type

struct FactorizationError <: Exception

Exception thrown if error occured during factorization.

VoronoiFVM.NewtonControl — Type

mutable struct NewtonControl

Control parameters for Newton method.

Newton's method solves $F(u)=0$ by the iterative procedure $u_{i+1}=u_{i} - d_i F'(u_i)^{-1}F(u_i)$ starting with some inital value $u_0$, where $d_i$ is the damping.

tol_absolute::Float64

Tolerance (in terms of norm of Newton update): terminate if $\Delta_i=||u_{i+1}-u_i||_\infty <$ tol_absolute.

Default value: 1.0e-10.

tol_relative::Float64

Tolerance (relative to the first residual): terminate if $\Delta_i/\Delta_0<$ tol_relative.

Default value: 1.0e-10.

tol_round::Float64

Tolerance for roundoff error detection: terminate if $|\;||u_{i+1}||_1 - ||u_{i}||_1\;|/ ||u_{i}||_1<$ tol_round occured max_round times in a row.

Default value: 1.0e-10.

tol_mono::Float64

Tolerance for monotonicity test: terminat with error if $\Delta_i/\Delta_{i-1}>$ 1/tol_mono.

Default value: 1.0e-3

damp_initial::Float64

Initial damping parameter $d_0$.

Default value: 1.0.

To handle convergence problems, set this to a value less than 1.

damp_growth::Float64

Damping parameter growth factor: $d_{i+1}=\min(d_i\cdot$ max_growth $,1)$

Default value: 1.2

Generally it should be set to a value between 1 and 2.

max_iterations::Int32

Maximum number of iterations.

Default value: 100

max_lureuse::Int32

Maximum number of reuses of lu factorization. It this value is 0, linear systems are solved by a sparse direct solver, and it's LU factorization is called in every Newton step.

Otherwise, a BICGstab iterative method is used for linear system solution with an LU factorization as preconditioner which is updated only every max_lureuse Newton step.

Default value: 0.

max_round::Int32

Maximum number of consecutive iterations within roundoff error tolerance

Default value: 1000 (effectively switching of this criterion).

tol_linear::Float64

Tolerance of iterative linear solver.

Default value: 1.0e-4.

verbose::Bool

Verbosity flag

Default value: false.

handle_exceptions::Bool

Handle exceptions in embed! and evolve! methods. If true, exceptions in Newton solves are catched, the embedding resp. time step is lowered, and solution is retried.

Default value: false

Δp::Float64

Initial parameter step for embed! method.

Default value: 1.0

Δp_max::Float64

Maximal parameter step for embed! method

Default value: 1.0

Δp_min::Float64

Minimal parameter step for embed! method.

Default value: 1.0e-3

Δt::Float64

Time step for evolve! method.

Default value: 0.1

Δt_max::Float64

Maximal time step for evolve! method.

Default value: 1

Δt_min::Float64

Minimal time step for evolve! method.

Default value: 1.0e-3

Δt_grow::Float64

Maximal step size growth for evolve! method.

Default: 1.2

Δu_opt::Float64

Optimal size of update for evolve! method. The algorithm keeps this value approximately constant.

Default: 0.1

force_first_step::Bool

force first timestep

edge_cutoff::Float64

Edge cutoff for rectangular triangles.

Default value: 0.0.

umfpack_pivot_tolerance::Float64

Pivot tolerance for umfpack

Default value: 0.1.

VoronoiFVM.NewtonControl — Method

NewtonControl(this::Any) -> Any

Default constructor

VoronoiFVM.TestFunctionFactory — Type

mutable struct TestFunctionFactory{Tv}

Data structure containing DenseSystem used to calculate test functions for boundary flux calculations.

system::VoronoiFVM.AbstractSystem{Tv, Ti, Tm} where {Tv, Ti<:Integer, Tm<:Integer}

Original system

tfsystem::VoronoiFVM.DenseSystem

Test function system

VoronoiFVM.TestFunctionFactory — Method

Constructor for TestFunctionFactory from System

VoronoiFVM.Dirichlet — Constant

Constant to be used as boundary condition factor to mark Dirichlet boundary conditons.

Methods

VoronoiFVM.System — Method

System(grid::Any, physics::VoronoiFVM.Physics; unknown_storage, matrixindextype) -> Any

Create Finite Volume System.

grid: 1D/2D/3D discretization grid

physics: Physics struct containing node and edge callbacks

unknown_storage: string or symbol:

:dense : solution vector is an nspecies x nnodes dense matrix

:sparse : solution vector is an nspecies x nnodes sparse matrix

VoronoiFVM.boundary_dirichlet! — Method

boundary_dirichlet!(this::VoronoiFVM.AbstractSystem, ispec::Integer, ibc::Integer, val::Any) -> Any

Set Dirichlet boundary conditon for species ispec at boundary ibc.

VoronoiFVM.boundary_neumann! — Method

boundary_neumann!(this::VoronoiFVM.AbstractSystem, ispec::Integer, ibc::Integer, val::Any) -> Any

Set Neumann boundary conditon for species ispec at boundary ibc.

VoronoiFVM.boundary_robin! — Method

boundary_robin!(this::VoronoiFVM.AbstractSystem, ispec::Integer, ibc::Integer, fac::Any, val::Any) -> Any

Set Robin boundary conditon for species ispec at boundary ibc.

VoronoiFVM.data — Method

data(this::VoronoiFVM.AbstractSystem) -> Any

Retrieve user data record.

VoronoiFVM.enable_boundary_species! — Method

enable_boundary_species!(this::VoronoiFVM.AbstractSystem, ispec::Integer, bregions::AbstractVector{T} where T)

Add species to a list of boundary regions. Species numbers for bulk and boundary species have to be distinct.

VoronoiFVM.enable_species! — Method

enable_species!(this::VoronoiFVM.AbstractSystem, ispec::Integer, regions::AbstractVector{T} where T)

Add species to a list of bulk regions. Species numbers for bulk and boundary species have to be distinct.

VoronoiFVM.is_boundary_species — Method

is_boundary_species(this::VoronoiFVM.AbstractSystem, ispec::Integer) -> Bool

Check if species number corresponds to boundary species.

VoronoiFVM.is_bulk_species — Method

is_bulk_species(this::VoronoiFVM.AbstractSystem, ispec::Integer) -> Bool

Check if species number corresponds bulk species.

VoronoiFVM.isdof — Method

isdof(this::VoronoiFVM.AbstractSystem, ispec::Any, inode::Any) -> Bool

Check if degree of freedom is defined.

VoronoiFVM.num_species — Method

num_species(a::AbstractArray) -> Any

Number of species (size of first dimension) of solution array.

VoronoiFVM.num_species — Method

num_species(this::VoronoiFVM.AbstractSystem) -> Any

Number of species in system

VoronoiFVM.update_grid! — Method

Update grid (e.g. after rescaling of coordinates).

VoronoiFVM.Grid — Function

Grid=ExtendableGrids.simplexgrid

Re-Export of ExtendableGrids.simplexgrid

VoronoiFVM.dof — Method

dof(a::Array{Tv, 2}, ispec::Integer, K::Integer) -> Any

Get degree of freedom number

VoronoiFVM.values — Method

Array of values in solution array.

Base.copy — Method

Create a copy of solution array

Base.getindex — Method

getindex(a::VoronoiFVM.SparseSolutionArray, ispec::Integer, inode::Integer) -> Any

Accessor for solution array.

Base.setindex! — Method

setindex!(a::VoronoiFVM.SparseSolutionArray, v::Any, ispec::Integer, inode::Integer) -> Union{Nothing, VoronoiFVM.SparseSolutionArray}

Accessor for solution array.

Base.similar — Method

Create a similar unintialized solution array

Base.size — Method

size(a::VoronoiFVM.SparseSolutionArray) -> Tuple{Int64, Int64}

Return size of solution array.

VoronoiFVM.dof — Method

dof(a, i, j)

Get number of degree of freedom. Return 0 if species is not defined in node.

VoronoiFVM.getdof — Method

getdof(a::VoronoiFVM.SparseSolutionArray, i::Integer) -> Any

Return value for degree of freedom.

VoronoiFVM.setdof! — Method

setdof!(a::VoronoiFVM.SparseSolutionArray, v::Any, i::Integer) -> Any

Set value for degree of freedom.

VoronoiFVM.values — Method

values(a::VoronoiFVM.SparseSolutionArray) -> Vector{Tv} where Tv

Array of values in solution array.

Base.reshape — Method

reshape(v, sys)

Reshape vector to fit as solution to system.

LinearAlgebra.norm — Method

VoronoiFVM.num_dof — Method

num_dof(this::VoronoiFVM.DenseSystem) -> Int64

Number of degrees of freedom for system.

VoronoiFVM.unknowns — Method

unknowns(Tu::Type, sys::VoronoiFVM.DenseSystem{Tv, Ti, Tm} where {Ti, Tm}; inival) -> Any

Create a solution vector for dense system. If inival is not specified, the entries of the returned vector are undefined.

VoronoiFVM.unknowns — Method

Create a solution vector for dense system. If inival is not specified, the entries of the returned vector are undefined.

Base.reshape — Method

reshape(v, sys)

Reshape vector to fit as solution to system.

VoronoiFVM.num_dof — Method

num_dof(this::VoronoiFVM.SparseSystem) -> Any

Number of degrees of freedom for system.

VoronoiFVM.unknowns — Method

unknowns(Tu, sys; inival)

Create a solution vector for sparse system with given type. If inival is not specified, the entries of the returned vector are undefined.

VoronoiFVM.unknowns — Method

Create a solution vector for sparse system. The entries of the returned vector are undefined.

VoronoiFVM.meas — Method

meas(edge::VoronoiFVM.Edge) -> Any

Calculate the length of an edge.

VoronoiFVM.num_species — Method

num_species(edge::VoronoiFVM.Edge) -> Any

Return number of species for edge

VoronoiFVM.unknowns — Method

unknowns(node::VoronoiFVM.BNode, u::Any) -> Any

Construct vector unknowns on bnode

VoronoiFVM.unknowns — Method

unknowns(node::VoronoiFVM.Node, u::Any) -> Any

Construct vector unknowns on node

VoronoiFVM.unknowns — Method

unknowns(edge::VoronoiFVM.Edge, u::Array{T, 1}, i::Any) -> VoronoiFVM.VectorUnknowns{_A} where _A

Construct vector unknowns on edge.

VoronoiFVM.unknowns — Method

unknowns(edge::VoronoiFVM.Edge, u::Array{T, 1}) -> VoronoiFVM.MatrixUnknowns{_A} where _A

Construct matrix of unknowns from edge - these can be used in flux functions with the v0.7.x and v0.8.x syntax to acces data.

VoronoiFVM.viewK — Method

viewK(edge::VoronoiFVM.Edge, u::Any) -> VoronoiFVM.VectorUnknowns{_A} where _A

Solution view on first edge node

VoronoiFVM.viewL — Method

viewL(edge::VoronoiFVM.Edge, u::Any) -> VoronoiFVM.VectorUnknowns{_A} where _A

Solution view on second edge node

VoronoiFVM.embed! — Method

function embed!(solution, inival, system; 
                control=NewtonControl(),
                pre=function(sol,p) end,
                post=function(sol,p) end
)

Solve stationary problem via parameter embedding, calling solve! for increasing values of the parameter p from interval (0,1). The user is responsible for the interpretation of the parameter. The optional pre() callback can be used to communicate its value to the system. The optional post() callback method can be used to perform various postprocessing steps.

If control.handle_error is true, solve! throws an error, and stepsize control.Δp is lowered, and solve! is called again with a smaller parameter value. If control.Δp<control.Δp_min, embed! is aborted with error.

VoronoiFVM.eval_and_assemble — Method

eval_and_assemble(system, U, UOld, F, tstep; edge_cutoff)

Main assembly method.

Evaluate solution with result in right hand side F and assemble matrix into system.matrix.

VoronoiFVM.evolve! — Method

function evolve!(solution, inival, system, times;
                 control=NewtonControl(), 
                 pre=function(sol,t) end,   
                 post=function(sol,oldsol, t, Δt) end,
                 sample=function(sol,t) end,
                 delta=(u,v,t, Δt)->norm(sys,u-v,Inf)
)

Use implicit Euler method + damped Newton's method to solve time dependent problem. Time step control is performed according to the data in control. All times in times are reached exactly.

Callbacks:

pre is invoked before each time step

post is invoked after each time step

sample is called for all times in times[2:end].

delta is used to control the time step.

VoronoiFVM.integrate — Method

integrate(system,F,U; boundary=false)    

Integrate node function (same signature as reaction or storage) F of solution vector over domain or boundary The result contains the integral for each species separately.

VoronoiFVM.solve! — Method

solve!(solution, inival, system; control=NewtonControl(),tstep=Inf, log=false)

Perform solution of stationary problem(if tstep==Inf) or one step of the implicit Euler method using Newton's method with inival as initial value. The method writes into solution.

It returns solution or, if log==true, a tuple of solution and a vector containing the residual history of Newton's method.

VoronoiFVM.solve — Method

function solve(inival, system, times;
               control=NewtonControl(), 
               pre=function(sol,t) end,   
               post=function(sol,oldsol, t, Δt) end,
               sample=function(sol,t) end,
               delta=(u,v,t, Δt)->norm(sys,u-v,Inf),
               store_all=true,
               in_memory=true
)

Use implicit Euler method + damped Newton's method to solve time dependent problem. Time step control is performed according to the data in control. All times in times are reached exactly.

Callbacks:

pre is invoked before each time step

post is invoked after each time step

sample is called for all times in times[2:end].

delta is used to control the time step.

If store_all==true, all timestep solutions are stored. Otherwise, only solutions for the elements of times are stored.

Returns a transient solution object sol containing stored solutions, see TransientSolution

VoronoiFVM.solve — Method

solve(inival, system; control=NewtonControl(), tstep=Inf, log=false)

Perform solution of stationary problem(if tstep==Inf) or one step of the implicit Euler method using Newton's method with inival as initial value. The method writes into solution.

It returns a solution array or, if log==true, a tuple of solution and a vector containing the residual history of Newton's method.

VoronoiFVM.value — Function

Extract value from dual number. Use to debug physics callbacks. Re-exported from ForwardDiff.jl

VoronoiFVM.fixed_timesteps! — Method

timesteps!(control,Δt; grow=1.0)

Modify control data such that the time steps are fixed to a geometric sequence such that Δtnew=Δtold*grow

VoronoiFVM.integrate — Method

integrate(this, tf, U, Uold, tstep)

Calculate test function integral for transient solution.

VoronoiFVM.integrate — Method

integrate(this, tf, U)

Calculate test function integral for steady state solution.

VoronoiFVM.integrate_stdy — Method

integrate_stdy(this, tf, U)

Steady state part of test function integral.

VoronoiFVM.integrate_tran — Method

Calculate transient part of test function integral.

VoronoiFVM.testfunction — Method

testfunction(factory::VoronoiFVM.TestFunctionFactory{Tv}, bc0::Any, bc1::Any) -> Any

Create testfunction which has Dirichlet zero boundary conditions for boundary regions in bc0 and Dirichlet one boundary conditions for boundary regions in bc1.