System

Types and Constants

VoronoiFVM.AbstractSystem — Type

abstract type AbstractSystem

Abstract type for finite volume system structure

VoronoiFVM.BNode — Type

mutable struct BNode{Tv}

Structure holding local boundary node information.

index::Int32

Index in grid

region::Int32

Boundary region number

coord::Array{Tv,1} where Tv

1D Array of node coordinates

nspec::Int64

Number of species defined in node

VoronoiFVM.DenseSystem — Type

mutable struct DenseSystem{Tv} <: VoronoiFVM.AbstractSystem{Tv}

Structure holding data for finite volume system solution. Information on species distribution is kept in dense matrices, and the solution array is of type Array{2}.

Unlike in the SparseSystem, the system matrix handles exactly those degrees of freedom which correspond to unknowns, and dummy degrees of freedom where unknowns are not defined. Handling of the sparse matrix structures for the bookeeping of the unknowns has less overhead, but additional dummy equations are added to the system matrix.

grid::VoronoiFVM.Grid

Grid

physics::VoronoiFVM.Physics

Physics data

boundary_values::Array{Tv,2} where Tv

Array of boundary values

boundary_factors::Array{Tv,2} where Tv

Array of boundary factors

region_species::Array{Int8,2}

Full matrix containing species numbers for inner regions

bregion_species::Array{Int8,2}

Full matrix containing species numbers for boundary regions

node_dof::Array{Int8,2}

Full matrix containing degree of freedom numbers for each node

matrix::ExtendableSparse.ExtendableSparseMatrix{Tv,Int64} where Tv

Jacobi matrix for nonlinear problem

species_homogeneous::Bool

Flag which says if the number of unknowns per node is constant

update::Array{Tv,2} where Tv

Solution vector holding Newton update

residual::Array{Tv,2} where Tv

Solution vector holding Newton residual

VoronoiFVM.DenseSystem — Method

DenseSystem(grid::VoronoiFVM.Grid, physics::VoronoiFVM.Physics) -> VoronoiFVM.DenseSystem{_1} where _1

Constructor for DenseSystem. physics provides some user data, maxspec is the maximum number of species.

VoronoiFVM.Edge — Type

mutable struct Edge{Tv}

Structure holding local edge information.

index::Int32

Index in grid

nodeK::Int32

Index of first node

nodeL::Int32

Index of second node

region::Int32

Inner region number corresponding to edge

coordK::Array{Tv,1} where Tv

1D Array of first node coordinates

coordL::Array{Tv,1} where Tv

1D Array of second node coordinates

nspec::Int64

Number of species defined in edge

icell::Int64

Number of discretization cell the edge is invoked from

VoronoiFVM.Node — Type

mutable struct Node{Tv}

Structure holding local node information.

index::Int32

Index in grid

region::Int32

Inner region number

coord::Array{Tv,1} where Tv

1D Array of node coordinates

nspec::Int64

Number of species defined in node

icell::Int64

Number of discretization cell the node is invoked from

VoronoiFVM.SparseSolutionArray — Type

struct SparseSolutionArray{Tv} <: AbstractArray{Tv,2}

Struct holding solution information for SparseSystem. Solution is stored in a sparse matrix structure.

This class plays well with the abstract array interface.

node_dof::SparseArrays.SparseMatrixCSC{Tv,Int32} where Tv

Sparse matrix holding actual data.

VoronoiFVM.SparseSystem — Type

mutable struct SparseSystem{Tv} <: VoronoiFVM.AbstractSystem{Tv}

Structure holding data for finite volume system solution. Information on species distribution is kept in sparse matrices, and the solution array is of type SparseSolutionArray, i.e. effectively it is a sparse matrix.

Unlike in the DenseSystem, the system matrix handles exactly those degrees of freedom which correspond to unknowns. However, handling of the sparse matrix structures for the bookeeping of the unknowns creates overhead.

grid::VoronoiFVM.Grid

Grid

physics::VoronoiFVM.Physics

Physics data

boundary_values::Array{Tv,2} where Tv

Array of boundary values

boundary_factors::Array{Tv,2} where Tv

Array of boundary factors

region_species::SparseArrays.SparseMatrixCSC{Int8,Int16}

Sparse matrix containing species numbers for inner regions

bregion_species::SparseArrays.SparseMatrixCSC{Int8,Int16}

Sparse matrix containing species numbers for boundary regions

node_dof::SparseArrays.SparseMatrixCSC{Int8,Int32}

Sparse matrix containing degree of freedom numbers for each node

matrix::ExtendableSparse.ExtendableSparseMatrix{Tv,Int64} where Tv

Jacobi matrix for nonlinear problem

species_homogeneous::Bool

Flag which says if the number of unknowns per node is constant

update::VoronoiFVM.SparseSolutionArray

Solution vector holding Newton update

residual::VoronoiFVM.SparseSolutionArray

Solution vector holding Newton residual

VoronoiFVM.SparseSystem — Method

SparseSystem(grid::VoronoiFVM.Grid, physics::VoronoiFVM.Physics) -> VoronoiFVM.SparseSystem{_1} where _1

Constructor for SparseSystem. physics provides some user data, maxspec is the maximum number of species.

VoronoiFVM.SubgridArrayView — Type

struct SubgridArrayView{Tv, Ta} <: AbstractArray{Tv,2}

Struct holding information for solution array view on subgrid

sysarray::Any

Original array

subgrid::VoronoiFVM.SubGrid

Subgrid for view

VoronoiFVM.AssemblyError — Type

struct AssemblyError <: Exception

Exception thrown if error occured during assembly (e.g. domain error)

VoronoiFVM.ConvergenceError — Type

struct ConvergenceError <: Exception

Exception thrown if Newton's method convergence fails.

VoronoiFVM.EmbeddingError — Type

struct EmbeddingError <: Exception

Exception thrown if embedding fails

VoronoiFVM.FactorizationError — Type

struct FactorizationError <: Exception

Exception thrown if error occured during factorization.

VoronoiFVM.NewtonControl — Type

mutable struct NewtonControl

Control parameter structure for Newton method.

tol_absolute::Float64

Tolerance (in terms of norm of Newton update)

tol_relative::Float64

Tolerance (relative to the first residual)

tol_round::Float64

Tolerance for roundoff error detection

damp_initial::Float64

Initial damping parameter

damp_growth::Float64

Damping parameter growth factor

max_iterations::Int32

Maximum number of iterations

max_lureuse::Int32

Maximum number of reuses of lu factorization

max_round::Int32

Maximum number of consecutive iterations within roundoff error tolerance

tol_linear::Float64

Tolerance of iterative linear solver

verbose::Bool

Verbosity flag

handle_exceptions::Bool

Handle exceptions

Δp::Float64

Parameter step for embedding

Δp_max::Float64

Maximal parameter step

Δp_min::Float64

Minimal parameter step

VoronoiFVM.NewtonControl — Method

NewtonControl(this::Any) -> Any

Default constructor

VoronoiFVM.TestFunctionFactory — Type

mutable struct TestFunctionFactory{Tv}

Data structure containing DenseSystem used to calculate test functions for boundary flux calculations.

system::VoronoiFVM.AbstractSystem{Tv} where Tv

Original system

tfsystem::VoronoiFVM.DenseSystem

Test function system

VoronoiFVM.TestFunctionFactory — Method

Constructor for TestFunctionFactory from System

VoronoiFVM.Dirichlet — Constant

Constant to be used as boundary condition factor to mark Dirichlet boundary conditons.

Methods

VoronoiFVM.data — Method

Retrieve user data record.

VoronoiFVM.dof — Method

dof(a::Tv, ispec::Tv, K::Tv)

Get number of degree of freedom.

VoronoiFVM.dof — Method

dof(a::Tv, i::Tv, j::Tv)

Get number of degree of freedom. Return 0 if species is not defined in node.

VoronoiFVM.edgelength — Method

Calculate the length of an edge.

VoronoiFVM.enable_boundary_species! — Method

enable_boundary_species!(this::VoronoiFVM.AbstractSystem, ispec::Integer, bregions::AbstractArray{T,1} where T)

Add species to a list of boundary regions. Species numbers for bulk and boundary species have to be distinct.

VoronoiFVM.enable_species! — Method

enable_species!(this::VoronoiFVM.AbstractSystem, ispec::Integer, regions::AbstractArray{T,1} where T)

Add species to a list of bulk regions. Species numbers for bulk and boundary species have to be distinct.

VoronoiFVM.getdof — Method

getdof(a::VoronoiFVM.SparseSolutionArray, i::Integer) -> Any

Return value for degree of freedom.

VoronoiFVM.num_dof — Method

num_dof(this::VoronoiFVM.DenseSystem) -> Int64

Number of degrees of freedom for system.

VoronoiFVM.num_dof — Method

num_dof(this::VoronoiFVM.SparseSystem) -> Int64

Number of degrees of freedom for system.

VoronoiFVM.num_nodes — Method

num_nodes(a::AbstractArray) -> Any

Number of nodes (size of second dimension) of solution array.

VoronoiFVM.setdof! — Method

setdof!(a::VoronoiFVM.SparseSolutionArray, v::Any, i::Integer) -> Any

Set value for degree of freedom.

VoronoiFVM.unknowns — Method

Create a solution vector for system.

VoronoiFVM.unknowns — Method

Create a solution vector for system.

VoronoiFVM.unknowns — Method

unknowns(Tu::Tv, sys::Tv) -> VoronoiFVM.SparseSolutionArray{_1} where _1

Create a solution vector for system.

VoronoiFVM.unknowns — Method

unknowns(Tu::Tv, sys::Tv) -> VoronoiFVM.SparseSolutionArray{_1} where _1

Create a solution vector for system with given type

VoronoiFVM.viewK — Method

viewK(nspec::Int64, u::AbstractArray) -> SubArray

Solution view on first edge node

VoronoiFVM.viewK — Method

viewK(edge::Tv, u::Tv)

Solution view on first edge node

VoronoiFVM.viewL — Method

viewL(nspec::Int64, u::AbstractArray) -> SubArray

Solution view on second edge node

VoronoiFVM.viewL — Method

Solution view on second edge node

VoronoiFVM.embed! — Method

Solution method for instance of abstract system.

Perform solution via parameter embedding, calling solve! for each value of the parameter p from interval (0,1). The user is responsible for the interpretation of the parameter. The optional pre() callback can be used to communicate its value to the system. The optionalpost() callback method can be used to perform various postprocessing steps.

If control.handle_error is true, solve! throws an error, and stepsize control.Δp is lowered, and solve! is called again with a smaller parameter value. If control.Δp<control.Δp_min, embed! is aborted with error.

VoronoiFVM.integrate — Method

integrate(this, F, U)

Integrate function F of solution vector over domain. The result contains the integral for each species separately.

VoronoiFVM.solve! — Method

solve!(solution::Tv, inival::Tv, this::Tv; control, tstep)

Solution method for instance of abstract system.

Perform solution of stationary system (if tstep==Inf) or one tine step of implicit Euler time step system using Newton's method with damping. Initial damping is chosen according control.damp_initial.

VoronoiFVM.value — Function

Extract value from dual number. Use to debug physics callbacks. Re-exported from ForwardDiff.jl

VoronoiFVM.integrate — Method

integrate(this::Tv, tf::Tv, U::Tv, Uold::Tv, tstep::Tv)

Calculate test function integral for transient solution.

VoronoiFVM.integrate — Method

integrate(this::Tv, tf::Tv, U::Tv)

Calculate test function integral for steady state solution.

VoronoiFVM.testfunction — Method

testfunction(factory::Tv, bc0::Tv, bc1::Tv)

Create testfunction which has Dirichlet zero boundary conditions for boundary regions in bc0 and Dirichlet one boundary conditions for boundary regions in bc1.

Base.copy — Method

Create a copy of solution array

Base.getindex — Method

getindex(a::VoronoiFVM.SparseSolutionArray, ispec::Integer, inode::Integer) -> Any

Accessor for solution array.

Base.getindex — Method

getindex(aview::VoronoiFVM.SubgridArrayView, ispec::Integer, inode::Integer) -> Any

Accessor method for subgrid array view.

Base.reshape — Method

reshape(v, sys)

Reshape vector to fit as solution to system.

Base.reshape — Method

reshape(v, sys)

Reshape vector to fit as solution to system.

Base.setindex! — Method

setindex!(a::VoronoiFVM.SparseSolutionArray, v::Any, ispec::Integer, inode::Integer) -> Union{Nothing, SparseSolutionArray}

Accessor for solution array.

Base.setindex! — Method

setindex!(aview::VoronoiFVM.SubgridArrayView, v::Any, ispec::Integer, inode::Integer) -> VoronoiFVM.SubgridArrayView

Accessor method for subgrid array view.

Base.size — Method

size(a::VoronoiFVM.SparseSolutionArray) -> Tuple{Int64,Int64}

Return size of solution array.

Base.size — Method

size(a::VoronoiFVM.SubgridArrayView) -> Tuple{Any,Int64}

Return size of solution array view.

Base.view — Method

view(a::Tv, sg::Tv) -> SubArray

Create a view of the solution array on a subgrid.

VoronoiFVM.is_boundary_species — Method

is_boundary_species(this::VoronoiFVM.AbstractSystem, ispec::Integer) -> Bool

Check if species number corresponds to boundary species.

VoronoiFVM.is_bulk_species — Method

is_bulk_species(this::VoronoiFVM.AbstractSystem, ispec::Integer) -> Bool

Check if species number corresponds bulk species.

VoronoiFVM.isdof — Method

isdof(this::VoronoiFVM.AbstractSystem, ispec::Any, inode::Any) -> Bool

Check if degree of freedom is defined.

VoronoiFVM.num_species — Method

num_species(a::Any) -> Any

Number of species (size of first dimension) of solution array.

VoronoiFVM.num_species — Method

Number of species in system

VoronoiFVM.num_species — Method

Return number of species for edge

VoronoiFVM.values — Method

values(a::Array) -> Array{T,1} where T

Array of values in solution array.

VoronoiFVM.values — Method

values(a::VoronoiFVM.SparseSolutionArray) -> Array{Tv,1} where Tv

Array of values in solution array.

VoronoiFVM.eval_and_assemble — Method

eval_and_assemble(system::Tv, U::Tv, UOld::Tv, F::Tv, tstep::Tv)

Wrapper for main assembly method.

Evaluate solution with result in right hand side F and assemble matrix into system.matrix.

VoronoiFVM.integrate_stdy — Method

integrate_stdy(this, tf, U)

Steady state part of test function integral.

VoronoiFVM.integrate_tran — Method

Calculate transient part of test function integral.